+++ .gitignore
@@ -1,4 +1,4 @@
 /dist
 /node_modules
 .nyc_output
-.env
+*.env
