--- list.ts
@@ -1,1236 +1,0 @@
-export const rss_resp = {
-  title: 'YTS RSS',
-  description: 'Most popular Torrents in the smallest file size RSS Feed',
-  link: 'https://yts.mx/',
-  image: '',
-  category: [],
-  items: [
-    {
-      title: 'The Search for General Tso (2014) [720p] [WEBRip] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/the-search-for-general-tso-2014"><img src="https://img.yts.mx/assets/images/movies/the_search_for_general_tso_2014/medium-cover.jpg" alt="The Search for General Tso (2014)" /></a><br />IMDB Rating: 6.9/10<br />Genre: Comedy / Documentary / History / Mystery<br />Size: 667.76 MB<br />Runtime: 1hr 11 min<br /><br />',
-      link: 'https://yts.mx/movies/the-search-for-general-tso-2014',
-      published: 1708133689000,
-      created: 1708133689000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/DD98BB9DE063BB5EBD8AE19E1EF2C571DF5165E2',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title:
-        'The Search for General Tso (2014) [1080p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/the-search-for-general-tso-2014"><img src="https://img.yts.mx/assets/images/movies/the_search_for_general_tso_2014/medium-cover.jpg" alt="The Search for General Tso (2014)" /></a><br />IMDB Rating: 6.9/10<br />Genre: Comedy / Documentary / History / Mystery<br />Size: 1.34 GB<br />Runtime: 1hr 11 min<br /><br />',
-      link: 'https://yts.mx/movies/the-search-for-general-tso-2014',
-      published: 1708136548000,
-      created: 1708136548000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/F388397957ABD2DD43B0F4E94B6638B30EA1979D',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'The Crow (1994) [720p] [BluRay] [YTS.MX]',
-      description:
-        "<a href=\"https://yts.mx/movies/the-crow-1994\"><img src=\"https://img.yts.mx/assets/images/movies/The_Crow_1994/medium-cover.jpg\" alt=\"The Crow (1994)\" /></a><br />IMDB Rating: 7.5/10<br />Genre: Action / Crime / Drama / Fantasy / Horror / Thriller<br />Size: 936.23 MB<br />Runtime: 1hr 42 min<br /><br />\"People once believed that when someone dies, a crow carries their soul to the land of the dead. But sometimes, something so bad happens that a terrible sadness is carried with it and the soul can't rest. Then sometimes, just sometimes, the crow can bring that soul back to put the wrong things right.\"October 30, Devil's Night, in Detroit. Sergeant Albrecht (Ernie Hudson) is at the scene of a crime where Shelly Webster (Sofia Shinas) has been beaten and raped and her fiancé, local musician and guitarist Eric Draven (Brandon Lee), has been murdered, having been stabbed, shot, and thrown out of their apartment window to the streets below. The couple had been planning their wedding for the next day; Halloween. A cop comments darkly, \"Who gets married on Halloween?\" Albrecht responds, \"Nobody.\"As he prepares to leave for the hospital with Shelly, Albrecht meets a young girl with a skateboard named Sarah (Rochelle Davis) who says she is a friend of Eric and Shelly and that they took care of her. Albrecht tells her that Shelly probably won't survive.One year later, a crow flies over the dark and rainy landscape of Detroit and lands on the headstone of Eric Draven. It taps at the stone with its beak, awakening Eric from the grave. He climbs out of the ground, trembling and wracked with convulsions. With some confusion, he stands and walks with the crow to his abandoned apartment, finding it in shambles. Memories of the night of his death come back to him as the gang responsible runs through the city, setting fire to it. The faces of the attackers stick in his mind and overcome with grief, he grabs the frame of a broken tympanum window, the glass cutting deep into his palms. To his astonishment, the blood from the wounds flows back into his skin and the wounds heal themselves. Bent on vengeance, he changes his clothes and paints his face white in parody of a porcelain harlequin mask he had given Shelly as a gift. He then draws black liner around his eyes and down his cheeks like tears and paints his lips black with Glasgow lines reaching out from the corners. Guided by the crow and sharing the crow's high perspective on the city, Eric sets out to avenge his and Shelly's murders.The first person he locates is Tin Tin (Laurence Mason), a thug obsessed with knives. Eric engages in hand-to-hand combat. He wins the fight by catching a knife and pinning Tin Tin to the wall with it. Eric confronts him about the murders the previous year and Tin Tin laughs in his face, mocking Shelly. Eric finishes him off by stabbing him in each vital organ (\"in alphabetical order\") with his own knives and takes his trench coat, leaving a crow-shaped bloodstain on the wall of the alley. Using Tin Tin's memories, he goes to the pawn shop where Tin Tin sold Shelly's engagement ring. Eric forces the owner, Gideon (Jon Polito), to show him where the ring is located and finds it mixed in a box with dozens of other rings. He lets Gideon live so that he can deliver a message of death to the rest of the gang before he douses the counters in gasoline, loads a shotgun with other rings of destroyed lives, and fires, igniting the shop in a massive explosion. Gideon escapes alive but badly burned.Eric locates the next thug on his list, Funboy (Michael Massee), getting high in an apartment with Sarah's drug-addicted mother, Darla (Anna Thomson). Funboy shoots Eric in his hand and Eric mocks him by pretending to scream in pain before showing the hole in his palm heal before his eyes. Frightened, Funboy is easily disarmed by Eric and is shot in the thigh before being injected with an overdose of morphine. Eric then confronts Darla but merely says that morphine is bad for her, forcing the drug out of her arm seemingly by magic, and tells her that she should return to her daughter who is surviving alone on the streets.Albrecht, meanwhile, has been following the deaths of the thugs around town. He doesn't share the rest of the department's view on the killings, thinking of them as more of a blessing than a curse. He even goes so far as to rescue Eric as he flees from choppers after his latest killing, though Eric is quick to disappear afterwards. Upon meeting again later on, Eric explains what he's doing and offers his best reason as to why he's come back. Albrecht tells him what he knows of Shelly's death and reveals that he stayed with her as she suffered through thirty hours of agony before dying. Eric touches Albrecht and receives from him all the pain felt by Shelly during that time.Gideon goes to Top Dollar (Michael Wincott), the leader of all gangs in Detroit, and warns him about Eric. Top Dollar responds by stabbing Gideon through the throat with a sword, sinking it to the hilt, as his lover/half sister/soothsayer, Myca (Bai Ling), looks on with glee. Around the same time, two of Top Dollar's thugs, T-Bird (David Patrick Kelly) and Skank (Angel David) stop by with the news about Tin-Tin's murder. Top Dollar shows only virtually no concern for Tin-Tin's demise and snorts cocaine from a large pile.T-Bird and Skank stop at a convenience store for supplies. Eric sneaks into the T-Bird's classic Ford Thunderbird and kidnaps him while Skank is in the store. Eric forces T-Bird to drive at high speed through the streets. Skank runs into the street, gets hit by a car & steals it to follow the pair; he watches as Eric ties T-Bird to the driver's seat of a car with explosives strapped to him & lets the car drive off the pier where it explodes midair and sinks into the harbor. When the police arrive, they discover a fiery symbol in the shape of a crow at the scene.Meanwhile, Darla tries to reconcile with Sarah by making her breakfast; though Sarah is skeptical at first of her mother's sudden change of heart, she accepts the effort and hugs her. Albrecht's \"superior\" chews him out just because he can & tells him he's suspended for misconduct.Sarah goes to Eric's apartment and speaks out to the shadows, feeling that he must be nearby. She tells him that she misses him and Shelly but, when no one responds, goes to leave and says that she figured he wouldn't care anymore. Eric speaks to her, appearing in the window, and tells her that, although they cannot be friends anymore, he still cares about her.Skank retreats to the gang hideout, rambling wildly about Eric; Sarah & Albrecht bond over their mutual experiences w/ Eric & discuss how he came back.At the gang hideout, Top Dollar, hardly concerned for T-Bird's fate, is discussing plans for criminal activity on Devil's Night with his other criminal associates. Eric crashes the meeting looking for Skank, and proceeds to mock everyone in the room. Dollar orders him killed and his associates all open fire, forcing Eric off the large conference table. He rises after a few moments and a massive gunfight ensues. Top Dollar escapes with Myca and his lieutenant, Grange (Tony Todd). Eric easily kills everyone in the room by gun, fists & swords, saving Skank for last, throwing him out the window. The police arrive to find Eric the last man standing; he escapes w/ Albrecht's help.Top Dollar, Grange, and Myca discuss Eric and his unique abilities. Myca reveals that Eric is invulnerable to harm, but only if his link to the real world, the crow, remains untouched. Realizing that if they kill the crow they can kill Eric and the three of them hatch a plan.Having killed the last of the four who had murdered him and Shelly, Eric returns to the cemetery and visits Shelly's grave. Sarah meets him there to say goodbye and he gives her Shelly's engagement ring as a necklace. Sarah leaves the cemetery only to be abducted by Grange and taken to a local church where Top Dollar and Myca are hiding out. Through the crow, Eric sees that Sarah is in danger and goes to rescue her. The crow flies inside ahead of him but is sniped by Grange. Eric, not realizing what wounding the crow has done for his own invulnerability, accosts Top Dollar, saying he'll let him leave if he gives Sarah to Eric. Top Dollar shoots Eric, now no longer invincible and beats him; Sarah is taken to the upper levels of the church. Albrecht, having gone to the church to pay his respects, realizes what's happening & comes in shooting. He finds Eric on the floor and helps him towards the spiral staircase in the back. The wounded crow is picked up by Myca, who hopes to harness its mystical powers. Albrecht manages to shoot and kill Grange, but is wounded himself. He stays below as Eric ascends the stairs. Myca meets him halfway up, holding the crow at knifepoint, but the crow lunges up and pecks her eyes out, causing her to fall from the stairs to her death.Eric makes it to the roof where Top Dollar is holding Sarah. He pushes her off the roof though she manages to hold onto some scaffolding as Top Dollar and Eric fight. Distracted for a moment by Sarah's cries for help, Eric is stabbed through the back by Top Dollar who admits to having everything to do with Eric and Shelly's deaths, claiming that nothing goes on in their town without his say-so. He tell Eric that he's enjoyed their sparring & that he'll miss him, but before he can kill him, Eric grabs him and bestows his gratitude for Top Dollar's crimes: all of the thirty hours of pain Shelly had to endure at once. In agonizing pain, Top Dollar falls off the roof and is impaled on a gargoyle.Eric rescues Sarah and brings her to Albrecht who is soon taken to an ambulance. Eric then disappears from sight and crawls to Shelly's grave before passing out. Her spirit comes to him, sans face paint, and they are finally reunited.Later, Sarah visits Eric's grave, which has closed up once again. The crow stands on his headstone, holding Shelly's ring that Sarah lost during her abduction. The crow drops it in Sarah's hand. As it flies off into the night, Sarah narrates that, though buildings burn and people die, real love is forever.The film closes with the words \"For Brandon and Eliza\", a dedication to both Brandon Lee, who was accidentally killed during filming in 1993, and Eliza Hutton, who was Brandon's partner, until his death in 1993.",
-      link: 'https://yts.mx/movies/the-crow-1994',
-      published: 1650134172000,
-      created: 1650134172000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/7FB1A9C726F3A996D172B216D7651239FE648B2A',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'The Crow (1994) [1080p] [BluRay] [5.1] [YTS.MX]',
-      description:
-        "<a href=\"https://yts.mx/movies/the-crow-1994\"><img src=\"https://img.yts.mx/assets/images/movies/The_Crow_1994/medium-cover.jpg\" alt=\"The Crow (1994)\" /></a><br />IMDB Rating: 7.5/10<br />Genre: Action / Crime / Drama / Fantasy / Horror / Thriller<br />Size: 1.88 GB<br />Runtime: 1hr 42 min<br /><br />\"People once believed that when someone dies, a crow carries their soul to the land of the dead. But sometimes, something so bad happens that a terrible sadness is carried with it and the soul can't rest. Then sometimes, just sometimes, the crow can bring that soul back to put the wrong things right.\"October 30, Devil's Night, in Detroit. Sergeant Albrecht (Ernie Hudson) is at the scene of a crime where Shelly Webster (Sofia Shinas) has been beaten and raped and her fiancé, local musician and guitarist Eric Draven (Brandon Lee), has been murdered, having been stabbed, shot, and thrown out of their apartment window to the streets below. The couple had been planning their wedding for the next day; Halloween. A cop comments darkly, \"Who gets married on Halloween?\" Albrecht responds, \"Nobody.\"As he prepares to leave for the hospital with Shelly, Albrecht meets a young girl with a skateboard named Sarah (Rochelle Davis) who says she is a friend of Eric and Shelly and that they took care of her. Albrecht tells her that Shelly probably won't survive.One year later, a crow flies over the dark and rainy landscape of Detroit and lands on the headstone of Eric Draven. It taps at the stone with its beak, awakening Eric from the grave. He climbs out of the ground, trembling and wracked with convulsions. With some confusion, he stands and walks with the crow to his abandoned apartment, finding it in shambles. Memories of the night of his death come back to him as the gang responsible runs through the city, setting fire to it. The faces of the attackers stick in his mind and overcome with grief, he grabs the frame of a broken tympanum window, the glass cutting deep into his palms. To his astonishment, the blood from the wounds flows back into his skin and the wounds heal themselves. Bent on vengeance, he changes his clothes and paints his face white in parody of a porcelain harlequin mask he had given Shelly as a gift. He then draws black liner around his eyes and down his cheeks like tears and paints his lips black with Glasgow lines reaching out from the corners. Guided by the crow and sharing the crow's high perspective on the city, Eric sets out to avenge his and Shelly's murders.The first person he locates is Tin Tin (Laurence Mason), a thug obsessed with knives. Eric engages in hand-to-hand combat. He wins the fight by catching a knife and pinning Tin Tin to the wall with it. Eric confronts him about the murders the previous year and Tin Tin laughs in his face, mocking Shelly. Eric finishes him off by stabbing him in each vital organ (\"in alphabetical order\") with his own knives and takes his trench coat, leaving a crow-shaped bloodstain on the wall of the alley. Using Tin Tin's memories, he goes to the pawn shop where Tin Tin sold Shelly's engagement ring. Eric forces the owner, Gideon (Jon Polito), to show him where the ring is located and finds it mixed in a box with dozens of other rings. He lets Gideon live so that he can deliver a message of death to the rest of the gang before he douses the counters in gasoline, loads a shotgun with other rings of destroyed lives, and fires, igniting the shop in a massive explosion. Gideon escapes alive but badly burned.Eric locates the next thug on his list, Funboy (Michael Massee), getting high in an apartment with Sarah's drug-addicted mother, Darla (Anna Thomson). Funboy shoots Eric in his hand and Eric mocks him by pretending to scream in pain before showing the hole in his palm heal before his eyes. Frightened, Funboy is easily disarmed by Eric and is shot in the thigh before being injected with an overdose of morphine. Eric then confronts Darla but merely says that morphine is bad for her, forcing the drug out of her arm seemingly by magic, and tells her that she should return to her daughter who is surviving alone on the streets.Albrecht, meanwhile, has been following the deaths of the thugs around town. He doesn't share the rest of the department's view on the killings, thinking of them as more of a blessing than a curse. He even goes so far as to rescue Eric as he flees from choppers after his latest killing, though Eric is quick to disappear afterwards. Upon meeting again later on, Eric explains what he's doing and offers his best reason as to why he's come back. Albrecht tells him what he knows of Shelly's death and reveals that he stayed with her as she suffered through thirty hours of agony before dying. Eric touches Albrecht and receives from him all the pain felt by Shelly during that time.Gideon goes to Top Dollar (Michael Wincott), the leader of all gangs in Detroit, and warns him about Eric. Top Dollar responds by stabbing Gideon through the throat with a sword, sinking it to the hilt, as his lover/half sister/soothsayer, Myca (Bai Ling), looks on with glee. Around the same time, two of Top Dollar's thugs, T-Bird (David Patrick Kelly) and Skank (Angel David) stop by with the news about Tin-Tin's murder. Top Dollar shows only virtually no concern for Tin-Tin's demise and snorts cocaine from a large pile.T-Bird and Skank stop at a convenience store for supplies. Eric sneaks into the T-Bird's classic Ford Thunderbird and kidnaps him while Skank is in the store. Eric forces T-Bird to drive at high speed through the streets. Skank runs into the street, gets hit by a car & steals it to follow the pair; he watches as Eric ties T-Bird to the driver's seat of a car with explosives strapped to him & lets the car drive off the pier where it explodes midair and sinks into the harbor. When the police arrive, they discover a fiery symbol in the shape of a crow at the scene.Meanwhile, Darla tries to reconcile with Sarah by making her breakfast; though Sarah is skeptical at first of her mother's sudden change of heart, she accepts the effort and hugs her. Albrecht's \"superior\" chews him out just because he can & tells him he's suspended for misconduct.Sarah goes to Eric's apartment and speaks out to the shadows, feeling that he must be nearby. She tells him that she misses him and Shelly but, when no one responds, goes to leave and says that she figured he wouldn't care anymore. Eric speaks to her, appearing in the window, and tells her that, although they cannot be friends anymore, he still cares about her.Skank retreats to the gang hideout, rambling wildly about Eric; Sarah & Albrecht bond over their mutual experiences w/ Eric & discuss how he came back.At the gang hideout, Top Dollar, hardly concerned for T-Bird's fate, is discussing plans for criminal activity on Devil's Night with his other criminal associates. Eric crashes the meeting looking for Skank, and proceeds to mock everyone in the room. Dollar orders him killed and his associates all open fire, forcing Eric off the large conference table. He rises after a few moments and a massive gunfight ensues. Top Dollar escapes with Myca and his lieutenant, Grange (Tony Todd). Eric easily kills everyone in the room by gun, fists & swords, saving Skank for last, throwing him out the window. The police arrive to find Eric the last man standing; he escapes w/ Albrecht's help.Top Dollar, Grange, and Myca discuss Eric and his unique abilities. Myca reveals that Eric is invulnerable to harm, but only if his link to the real world, the crow, remains untouched. Realizing that if they kill the crow they can kill Eric and the three of them hatch a plan.Having killed the last of the four who had murdered him and Shelly, Eric returns to the cemetery and visits Shelly's grave. Sarah meets him there to say goodbye and he gives her Shelly's engagement ring as a necklace. Sarah leaves the cemetery only to be abducted by Grange and taken to a local church where Top Dollar and Myca are hiding out. Through the crow, Eric sees that Sarah is in danger and goes to rescue her. The crow flies inside ahead of him but is sniped by Grange. Eric, not realizing what wounding the crow has done for his own invulnerability, accosts Top Dollar, saying he'll let him leave if he gives Sarah to Eric. Top Dollar shoots Eric, now no longer invincible and beats him; Sarah is taken to the upper levels of the church. Albrecht, having gone to the church to pay his respects, realizes what's happening & comes in shooting. He finds Eric on the floor and helps him towards the spiral staircase in the back. The wounded crow is picked up by Myca, who hopes to harness its mystical powers. Albrecht manages to shoot and kill Grange, but is wounded himself. He stays below as Eric ascends the stairs. Myca meets him halfway up, holding the crow at knifepoint, but the crow lunges up and pecks her eyes out, causing her to fall from the stairs to her death.Eric makes it to the roof where Top Dollar is holding Sarah. He pushes her off the roof though she manages to hold onto some scaffolding as Top Dollar and Eric fight. Distracted for a moment by Sarah's cries for help, Eric is stabbed through the back by Top Dollar who admits to having everything to do with Eric and Shelly's deaths, claiming that nothing goes on in their town without his say-so. He tell Eric that he's enjoyed their sparring & that he'll miss him, but before he can kill him, Eric grabs him and bestows his gratitude for Top Dollar's crimes: all of the thirty hours of pain Shelly had to endure at once. In agonizing pain, Top Dollar falls off the roof and is impaled on a gargoyle.Eric rescues Sarah and brings her to Albrecht who is soon taken to an ambulance. Eric then disappears from sight and crawls to Shelly's grave before passing out. Her spirit comes to him, sans face paint, and they are finally reunited.Later, Sarah visits Eric's grave, which has closed up once again. The crow stands on his headstone, holding Shelly's ring that Sarah lost during her abduction. The crow drops it in Sarah's hand. As it flies off into the night, Sarah narrates that, though buildings burn and people die, real love is forever.The film closes with the words \"For Brandon and Eliza\", a dedication to both Brandon Lee, who was accidentally killed during filming in 1993, and Eliza Hutton, who was Brandon's partner, until his death in 1993.",
-      link: 'https://yts.mx/movies/the-crow-1994',
-      published: 1650138715000,
-      created: 1650138715000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/58E6BCB2321181B92DBD7565C1F08D74B4DF2D6B',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'The Crow (1994) [2160p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]',
-      description:
-        "<a href=\"https://yts.mx/movies/the-crow-1994\"><img src=\"https://img.yts.mx/assets/images/movies/The_Crow_1994/medium-cover.jpg\" alt=\"The Crow (1994)\" /></a><br />IMDB Rating: 7.5/10<br />Genre: Action / Crime / Drama / Fantasy / Horror / Thriller<br />Size: 4.53 GB<br />Runtime: 1hr 42 min<br /><br />\"People once believed that when someone dies, a crow carries their soul to the land of the dead. But sometimes, something so bad happens that a terrible sadness is carried with it and the soul can't rest. Then sometimes, just sometimes, the crow can bring that soul back to put the wrong things right.\"October 30, Devil's Night, in Detroit. Sergeant Albrecht (Ernie Hudson) is at the scene of a crime where Shelly Webster (Sofia Shinas) has been beaten and raped and her fiancé, local musician and guitarist Eric Draven (Brandon Lee), has been murdered, having been stabbed, shot, and thrown out of their apartment window to the streets below. The couple had been planning their wedding for the next day; Halloween. A cop comments darkly, \"Who gets married on Halloween?\" Albrecht responds, \"Nobody.\"As he prepares to leave for the hospital with Shelly, Albrecht meets a young girl with a skateboard named Sarah (Rochelle Davis) who says she is a friend of Eric and Shelly and that they took care of her. Albrecht tells her that Shelly probably won't survive.One year later, a crow flies over the dark and rainy landscape of Detroit and lands on the headstone of Eric Draven. It taps at the stone with its beak, awakening Eric from the grave. He climbs out of the ground, trembling and wracked with convulsions. With some confusion, he stands and walks with the crow to his abandoned apartment, finding it in shambles. Memories of the night of his death come back to him as the gang responsible runs through the city, setting fire to it. The faces of the attackers stick in his mind and overcome with grief, he grabs the frame of a broken tympanum window, the glass cutting deep into his palms. To his astonishment, the blood from the wounds flows back into his skin and the wounds heal themselves. Bent on vengeance, he changes his clothes and paints his face white in parody of a porcelain harlequin mask he had given Shelly as a gift. He then draws black liner around his eyes and down his cheeks like tears and paints his lips black with Glasgow lines reaching out from the corners. Guided by the crow and sharing the crow's high perspective on the city, Eric sets out to avenge his and Shelly's murders.The first person he locates is Tin Tin (Laurence Mason), a thug obsessed with knives. Eric engages in hand-to-hand combat. He wins the fight by catching a knife and pinning Tin Tin to the wall with it. Eric confronts him about the murders the previous year and Tin Tin laughs in his face, mocking Shelly. Eric finishes him off by stabbing him in each vital organ (\"in alphabetical order\") with his own knives and takes his trench coat, leaving a crow-shaped bloodstain on the wall of the alley. Using Tin Tin's memories, he goes to the pawn shop where Tin Tin sold Shelly's engagement ring. Eric forces the owner, Gideon (Jon Polito), to show him where the ring is located and finds it mixed in a box with dozens of other rings. He lets Gideon live so that he can deliver a message of death to the rest of the gang before he douses the counters in gasoline, loads a shotgun with other rings of destroyed lives, and fires, igniting the shop in a massive explosion. Gideon escapes alive but badly burned.Eric locates the next thug on his list, Funboy (Michael Massee), getting high in an apartment with Sarah's drug-addicted mother, Darla (Anna Thomson). Funboy shoots Eric in his hand and Eric mocks him by pretending to scream in pain before showing the hole in his palm heal before his eyes. Frightened, Funboy is easily disarmed by Eric and is shot in the thigh before being injected with an overdose of morphine. Eric then confronts Darla but merely says that morphine is bad for her, forcing the drug out of her arm seemingly by magic, and tells her that she should return to her daughter who is surviving alone on the streets.Albrecht, meanwhile, has been following the deaths of the thugs around town. He doesn't share the rest of the department's view on the killings, thinking of them as more of a blessing than a curse. He even goes so far as to rescue Eric as he flees from choppers after his latest killing, though Eric is quick to disappear afterwards. Upon meeting again later on, Eric explains what he's doing and offers his best reason as to why he's come back. Albrecht tells him what he knows of Shelly's death and reveals that he stayed with her as she suffered through thirty hours of agony before dying. Eric touches Albrecht and receives from him all the pain felt by Shelly during that time.Gideon goes to Top Dollar (Michael Wincott), the leader of all gangs in Detroit, and warns him about Eric. Top Dollar responds by stabbing Gideon through the throat with a sword, sinking it to the hilt, as his lover/half sister/soothsayer, Myca (Bai Ling), looks on with glee. Around the same time, two of Top Dollar's thugs, T-Bird (David Patrick Kelly) and Skank (Angel David) stop by with the news about Tin-Tin's murder. Top Dollar shows only virtually no concern for Tin-Tin's demise and snorts cocaine from a large pile.T-Bird and Skank stop at a convenience store for supplies. Eric sneaks into the T-Bird's classic Ford Thunderbird and kidnaps him while Skank is in the store. Eric forces T-Bird to drive at high speed through the streets. Skank runs into the street, gets hit by a car & steals it to follow the pair; he watches as Eric ties T-Bird to the driver's seat of a car with explosives strapped to him & lets the car drive off the pier where it explodes midair and sinks into the harbor. When the police arrive, they discover a fiery symbol in the shape of a crow at the scene.Meanwhile, Darla tries to reconcile with Sarah by making her breakfast; though Sarah is skeptical at first of her mother's sudden change of heart, she accepts the effort and hugs her. Albrecht's \"superior\" chews him out just because he can & tells him he's suspended for misconduct.Sarah goes to Eric's apartment and speaks out to the shadows, feeling that he must be nearby. She tells him that she misses him and Shelly but, when no one responds, goes to leave and says that she figured he wouldn't care anymore. Eric speaks to her, appearing in the window, and tells her that, although they cannot be friends anymore, he still cares about her.Skank retreats to the gang hideout, rambling wildly about Eric; Sarah & Albrecht bond over their mutual experiences w/ Eric & discuss how he came back.At the gang hideout, Top Dollar, hardly concerned for T-Bird's fate, is discussing plans for criminal activity on Devil's Night with his other criminal associates. Eric crashes the meeting looking for Skank, and proceeds to mock everyone in the room. Dollar orders him killed and his associates all open fire, forcing Eric off the large conference table. He rises after a few moments and a massive gunfight ensues. Top Dollar escapes with Myca and his lieutenant, Grange (Tony Todd). Eric easily kills everyone in the room by gun, fists & swords, saving Skank for last, throwing him out the window. The police arrive to find Eric the last man standing; he escapes w/ Albrecht's help.Top Dollar, Grange, and Myca discuss Eric and his unique abilities. Myca reveals that Eric is invulnerable to harm, but only if his link to the real world, the crow, remains untouched. Realizing that if they kill the crow they can kill Eric and the three of them hatch a plan.Having killed the last of the four who had murdered him and Shelly, Eric returns to the cemetery and visits Shelly's grave. Sarah meets him there to say goodbye and he gives her Shelly's engagement ring as a necklace. Sarah leaves the cemetery only to be abducted by Grange and taken to a local church where Top Dollar and Myca are hiding out. Through the crow, Eric sees that Sarah is in danger and goes to rescue her. The crow flies inside ahead of him but is sniped by Grange. Eric, not realizing what wounding the crow has done for his own invulnerability, accosts Top Dollar, saying he'll let him leave if he gives Sarah to Eric. Top Dollar shoots Eric, now no longer invincible and beats him; Sarah is taken to the upper levels of the church. Albrecht, having gone to the church to pay his respects, realizes what's happening & comes in shooting. He finds Eric on the floor and helps him towards the spiral staircase in the back. The wounded crow is picked up by Myca, who hopes to harness its mystical powers. Albrecht manages to shoot and kill Grange, but is wounded himself. He stays below as Eric ascends the stairs. Myca meets him halfway up, holding the crow at knifepoint, but the crow lunges up and pecks her eyes out, causing her to fall from the stairs to her death.Eric makes it to the roof where Top Dollar is holding Sarah. He pushes her off the roof though she manages to hold onto some scaffolding as Top Dollar and Eric fight. Distracted for a moment by Sarah's cries for help, Eric is stabbed through the back by Top Dollar who admits to having everything to do with Eric and Shelly's deaths, claiming that nothing goes on in their town without his say-so. He tell Eric that he's enjoyed their sparring & that he'll miss him, but before he can kill him, Eric grabs him and bestows his gratitude for Top Dollar's crimes: all of the thirty hours of pain Shelly had to endure at once. In agonizing pain, Top Dollar falls off the roof and is impaled on a gargoyle.Eric rescues Sarah and brings her to Albrecht who is soon taken to an ambulance. Eric then disappears from sight and crawls to Shelly's grave before passing out. Her spirit comes to him, sans face paint, and they are finally reunited.Later, Sarah visits Eric's grave, which has closed up once again. The crow stands on his headstone, holding Shelly's ring that Sarah lost during her abduction. The crow drops it in Sarah's hand. As it flies off into the night, Sarah narrates that, though buildings burn and people die, real love is forever.The film closes with the words \"For Brandon and Eliza\", a dedication to both Brandon Lee, who was accidentally killed during filming in 1993, and Eliza Hutton, who was Brandon's partner, until his death in 1993.",
-      link: 'https://yts.mx/movies/the-crow-1994',
-      published: 1708133420000,
-      created: 1708133420000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/DA28E9B261AE01174610C6704CF9DD306F598CE6',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: "Dounia et la princesse d'Alep (2022) [720p] [WEBRip] [YTS.MX]",
-      description:
-        '<a href="https://yts.mx/movies/dounia-et-la-princesse-dalep-2022"><img src="https://img.yts.mx/assets/images/movies/dounia_et_la_princesse_dalep_2022/medium-cover.jpg" alt="Dounia et la princesse d\'Alep (2022)" /></a><br />IMDB Rating: 6.9/10<br />Genre: Animation / Family<br />Size: 649.52 MB<br />Runtime: 1hr 12 min<br /><br />',
-      link: 'https://yts.mx/movies/dounia-et-la-princesse-dalep-2022',
-      published: 1708129684000,
-      created: 1708129684000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/5E99C84DE82BB63005F38A8B6E7678A47D925B74',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title:
-        "Dounia et la princesse d'Alep (2022) [1080p] [WEBRip] [5.1] [YTS.MX]",
-      description:
-        '<a href="https://yts.mx/movies/dounia-et-la-princesse-dalep-2022"><img src="https://img.yts.mx/assets/images/movies/dounia_et_la_princesse_dalep_2022/medium-cover.jpg" alt="Dounia et la princesse d\'Alep (2022)" /></a><br />IMDB Rating: 6.9/10<br />Genre: Animation / Family<br />Size: 1.29 GB<br />Runtime: 1hr 12 min<br /><br />',
-      link: 'https://yts.mx/movies/dounia-et-la-princesse-dalep-2022',
-      published: 1708132070000,
-      created: 1708132070000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/16502AD40B2B5D79A956945BA6929B190B2A5220',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'Lights Out (2024) [720p] [WEBRip] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/lights-out-2024"><img src="https://img.yts.mx/assets/images/movies/lights_out_2024/medium-cover.jpg" alt="Lights Out (2024)" /></a><br />IMDB Rating: 5.4/10<br />Genre: Action / Thriller<br />Size: 828.59 MB<br />Runtime: 1hr 30 min<br /><br />A homeless veteran, Michael "Duffy" Duffield (Grillo), is always on the move, until he meets a talkative Ex-Con, Max Bomer (Phifer). Max notices Duffy\'s skills after he gets into a bar fight and offers him a well-paying job. Duffy agrees and discovers that the "job" is competing in underground fight clubs. The pair make a ton of money on the first fight and decide to travel to LA so Duffy can atone for his past and Max can pay back a crime boss, Sage Parker (Mulroney). They stay with Rachel, Max\'s sister, and Duffy quickly forms a connection with her. Duffy enters Sage\'s fight club and eventually wins, but it leads to unwanted attention. Duffy gets tied up in the crime world and offered jobs he can\'t refuse, including one with Sage\'s partner and Police officer, Ellen Ridgway. The deeper he goes in this world, the more deadly it gets. Duffy is a skilled fighter, but will he and Max make it out alive?—The Exchange company',
-      link: 'https://yts.mx/movies/lights-out-2024',
-      published: 1708122498000,
-      created: 1708122498000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/1353AA70EFB21BD212660AF3676308B07E283D58',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'Lights Out (2024) [1080p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/lights-out-2024"><img src="https://img.yts.mx/assets/images/movies/lights_out_2024/medium-cover.jpg" alt="Lights Out (2024)" /></a><br />IMDB Rating: 5.4/10<br />Genre: Action / Thriller<br />Size: 1.66 GB<br />Runtime: 1hr 30 min<br /><br />A homeless veteran, Michael "Duffy" Duffield (Grillo), is always on the move, until he meets a talkative Ex-Con, Max Bomer (Phifer). Max notices Duffy\'s skills after he gets into a bar fight and offers him a well-paying job. Duffy agrees and discovers that the "job" is competing in underground fight clubs. The pair make a ton of money on the first fight and decide to travel to LA so Duffy can atone for his past and Max can pay back a crime boss, Sage Parker (Mulroney). They stay with Rachel, Max\'s sister, and Duffy quickly forms a connection with her. Duffy enters Sage\'s fight club and eventually wins, but it leads to unwanted attention. Duffy gets tied up in the crime world and offered jobs he can\'t refuse, including one with Sage\'s partner and Police officer, Ellen Ridgway. The deeper he goes in this world, the more deadly it gets. Duffy is a skilled fighter, but will he and Max make it out alive?—The Exchange company',
-      link: 'https://yts.mx/movies/lights-out-2024',
-      published: 1708126104000,
-      created: 1708126104000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/2C882BEE2E0985DF54C6DC82C55AEDA67E243621',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'The Heartbreak Agency (2024) [720p] [WEBRip] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/the-heartbreak-agency-2024"><img src="https://img.yts.mx/assets/images/movies/the_heartbreak_agency_2024/medium-cover.jpg" alt="The Heartbreak Agency (2024)" /></a><br />IMDB Rating: 5.7/10<br />Genre: Comedy / Romance<br />Size: 889.79 MB<br />Runtime: 1hr 34 min<br /><br />In a world where lovesickness is seen as outdated and superfluous, a young journalist decides to expose the founder of a lovesickness agency as a fraud. He firmly believes that no one can suffer from lovesickness these days and that the agency is just a clever hoax.',
-      link: 'https://yts.mx/movies/the-heartbreak-agency-2024',
-      published: 1708118424000,
-      created: 1708118424000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/502C7D352C70A3F617F00A214510B21E4F390F28',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'The Heartbreak Agency (2024) [1080p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/the-heartbreak-agency-2024"><img src="https://img.yts.mx/assets/images/movies/the_heartbreak_agency_2024/medium-cover.jpg" alt="The Heartbreak Agency (2024)" /></a><br />IMDB Rating: 5.7/10<br />Genre: Comedy / Romance<br />Size: 1.78 GB<br />Runtime: 1hr 34 min<br /><br />In a world where lovesickness is seen as outdated and superfluous, a young journalist decides to expose the founder of a lovesickness agency as a fraud. He firmly believes that no one can suffer from lovesickness these days and that the agency is just a clever hoax.',
-      link: 'https://yts.mx/movies/the-heartbreak-agency-2024',
-      published: 1708121964000,
-      created: 1708121964000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/3AEAA53F0F5EBBB84BC9788D88D8B80D247EE993',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'A Soweto Love Story (2024) [720p] [WEBRip] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/a-soweto-love-story-2024"><img src="https://img.yts.mx/assets/images/movies/a_soweto_love_story_2024/medium-cover.jpg" alt="A Soweto Love Story (2024)" /></a><br />IMDB Rating: 4.8/10<br />Genre: Comedy / Romance<br />Size: 939.93 MB<br />Runtime: 1hr 41 min<br /><br />When her three grown sons show no signs of settling down and starting their own families, a mother uses the coveted family home to entice them into doing the right thing- the first to get married before the New Year wins the home. Her actions set into motion a series of humorous events as the brothers race to find brides.',
-      link: 'https://yts.mx/movies/a-soweto-love-story-2024',
-      published: 1708115422000,
-      created: 1708115422000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/5F4F57F68136592EE74D4571D467A0BCF0CC7799',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'A Soweto Love Story (2024) [1080p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/a-soweto-love-story-2024"><img src="https://img.yts.mx/assets/images/movies/a_soweto_love_story_2024/medium-cover.jpg" alt="A Soweto Love Story (2024)" /></a><br />IMDB Rating: 4.8/10<br />Genre: Comedy / Romance<br />Size: 1.88 GB<br />Runtime: 1hr 41 min<br /><br />When her three grown sons show no signs of settling down and starting their own families, a mother uses the coveted family home to entice them into doing the right thing- the first to get married before the New Year wins the home. Her actions set into motion a series of humorous events as the brothers race to find brides.',
-      link: 'https://yts.mx/movies/a-soweto-love-story-2024',
-      published: 1708119113000,
-      created: 1708119113000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/4C06C49CF4FD8BFD3C37533F63D8D0695A682988',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'Land of Bad (2024) [720p] [WEBRip] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/land-of-bad-2024"><img src="https://img.yts.mx/assets/images/movies/land_of_bad_2024/medium-cover.jpg" alt="Land of Bad (2024)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Action / Thriller<br />Size: 1.02 GB<br />Runtime: 1hr 50 min<br /><br />',
-      link: 'https://yts.mx/movies/land-of-bad-2024',
-      published: 1708076506000,
-      created: 1708076506000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/413E1AE0315078519C27D0E467CE9B80C96F8AA7',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'Land of Bad (2024) [1080p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/land-of-bad-2024"><img src="https://img.yts.mx/assets/images/movies/land_of_bad_2024/medium-cover.jpg" alt="Land of Bad (2024)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Action / Thriller<br />Size: 2.09 GB<br />Runtime: 1hr 50 min<br /><br />',
-      link: 'https://yts.mx/movies/land-of-bad-2024',
-      published: 1708080765000,
-      created: 1708080765000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/9E0B1716A04EA0E43765350AB4A9FC2038064947',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title:
-        'Land of Bad (2024) [1080p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/land-of-bad-2024"><img src="https://img.yts.mx/assets/images/movies/land_of_bad_2024/medium-cover.jpg" alt="Land of Bad (2024)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Action / Thriller<br />Size: 1.89 GB<br />Runtime: 1hr 50 min<br /><br />',
-      link: 'https://yts.mx/movies/land-of-bad-2024',
-      published: 1708116926000,
-      created: 1708116926000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/FB6869A5C96490E66869ADB4D3822A911324D31A',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'The Group (2022) [720p] [WEBRip] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/the-group-2022"><img src="https://img.yts.mx/assets/images/movies/the_group_2022/medium-cover.jpg" alt="The Group (2022)" /></a><br />IMDB Rating: 7.3/10<br />Genre: Crime / Horror / Thriller<br />Size: 681.18 MB<br />Runtime: 1hr 11 min<br /><br />',
-      link: 'https://yts.mx/movies/the-group-2022',
-      published: 1708114078000,
-      created: 1708114078000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/1AD4AB2B385833B6B574C31F6BDDEE121AB390D9',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'The Group (2022) [1080p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/the-group-2022"><img src="https://img.yts.mx/assets/images/movies/the_group_2022/medium-cover.jpg" alt="The Group (2022)" /></a><br />IMDB Rating: 7.3/10<br />Genre: Crime / Horror / Thriller<br />Size: 1.37 GB<br />Runtime: 1hr 11 min<br /><br />',
-      link: 'https://yts.mx/movies/the-group-2022',
-      published: 1708116368000,
-      created: 1708116368000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/F813CDD7BAC79FD4D12C9F578CD00A1038ACAB7D',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title:
-        'In the Land of Saints and Sinners (2023) [720p] [BluRay] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/in-the-land-of-saints-and-sinners-2023"><img src="https://img.yts.mx/assets/images/movies/in_the_land_of_saints_and_sinners_2023/medium-cover.jpg" alt="In the Land of Saints and Sinners (2023)" /></a><br />IMDB Rating: 6.2/10<br />Genre: Action / Crime / Thriller<br />Size: 978.12 MB<br />Runtime: 1hr 46 min<br /><br />',
-      link: 'https://yts.mx/movies/in-the-land-of-saints-and-sinners-2023',
-      published: 1708109487000,
-      created: 1708109487000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/46EEFFC4B098511E242AB250A8557662D555C315',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title:
-        'In the Land of Saints and Sinners (2023) [1080p] [BluRay] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/in-the-land-of-saints-and-sinners-2023"><img src="https://img.yts.mx/assets/images/movies/in_the_land_of_saints_and_sinners_2023/medium-cover.jpg" alt="In the Land of Saints and Sinners (2023)" /></a><br />IMDB Rating: 6.2/10<br />Genre: Action / Crime / Thriller<br />Size: 1.96 GB<br />Runtime: 1hr 46 min<br /><br />',
-      link: 'https://yts.mx/movies/in-the-land-of-saints-and-sinners-2023',
-      published: 1708113278000,
-      created: 1708113278000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/8F78DAC45FBCBAFDD19DE7A7532AB5002C71B4AE',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title:
-        'In the Land of Saints and Sinners (2023) [720p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/in-the-land-of-saints-and-sinners-2023"><img src="https://img.yts.mx/assets/images/movies/in_the_land_of_saints_and_sinners_2023/medium-cover.jpg" alt="In the Land of Saints and Sinners (2023)" /></a><br />IMDB Rating: 6.2/10<br />Genre: Action / Crime / Thriller<br />Size: 936.5 MB<br />Runtime: 1hr 46 min<br /><br />',
-      link: 'https://yts.mx/movies/in-the-land-of-saints-and-sinners-2023',
-      published: 1701114038000,
-      created: 1701114038000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/E3CB3FFCDF47253174F256363C793E6F16967CE4',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title:
-        'In the Land of Saints and Sinners (2023) [1080p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/in-the-land-of-saints-and-sinners-2023"><img src="https://img.yts.mx/assets/images/movies/in_the_land_of_saints_and_sinners_2023/medium-cover.jpg" alt="In the Land of Saints and Sinners (2023)" /></a><br />IMDB Rating: 6.2/10<br />Genre: Action / Crime / Thriller<br />Size: 1.88 GB<br />Runtime: 1hr 46 min<br /><br />',
-      link: 'https://yts.mx/movies/in-the-land-of-saints-and-sinners-2023',
-      published: 1701117604000,
-      created: 1701117604000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/4D6E1FDC8E2CA99B6F5CA57DE5A9592E60BC5BBA',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'The Good Mother (2023) [720p] [BluRay] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/the-good-mother-2023"><img src="https://img.yts.mx/assets/images/movies/the_good_mother_2023/medium-cover.jpg" alt="The Good Mother (2023)" /></a><br />IMDB Rating: 4.7/10<br />Genre: Action / Crime / Drama / Mystery / Thriller<br />Size: 831.9 MB<br />Runtime: 1hr 30 min<br /><br />From the three-time Academy Award®-nominated producer of The Wolf of Wall Street, The Good Mother follows journalist Marissa Bennings who, after the murder of her estranged son, forms an unlikely alliance with his pregnant girlfriend Paige to track down the killers. Together they confront a world of corruption and drugs in the seedy underbelly of a small city in upstate New York. As they get closer to the truth, they unearth an even darker secret. Starring two-time Academy Award®-winner Hilary Swank, Olivia Cooke and Jack Reynor.',
-      link: 'https://yts.mx/movies/the-good-mother-2023',
-      published: 1708101674000,
-      created: 1708101674000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/DAA16CB530743684156E66B9E5C3DAC1E6F95683',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'The Good Mother (2023) [1080p] [BluRay] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/the-good-mother-2023"><img src="https://img.yts.mx/assets/images/movies/the_good_mother_2023/medium-cover.jpg" alt="The Good Mother (2023)" /></a><br />IMDB Rating: 4.7/10<br />Genre: Action / Crime / Drama / Mystery / Thriller<br />Size: 1.67 GB<br />Runtime: 1hr 30 min<br /><br />From the three-time Academy Award®-nominated producer of The Wolf of Wall Street, The Good Mother follows journalist Marissa Bennings who, after the murder of her estranged son, forms an unlikely alliance with his pregnant girlfriend Paige to track down the killers. Together they confront a world of corruption and drugs in the seedy underbelly of a small city in upstate New York. As they get closer to the truth, they unearth an even darker secret. Starring two-time Academy Award®-winner Hilary Swank, Olivia Cooke and Jack Reynor.',
-      link: 'https://yts.mx/movies/the-good-mother-2023',
-      published: 1708105303000,
-      created: 1708105303000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/F166B14CAF0C14E6175F351AFA72576F89651F6B',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'The Good Mother (2023) [720p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/the-good-mother-2023"><img src="https://img.yts.mx/assets/images/movies/the_good_mother_2023/medium-cover.jpg" alt="The Good Mother (2023)" /></a><br />IMDB Rating: 4.7/10<br />Genre: Action / Crime / Drama / Mystery / Thriller<br />Size: 831.11 MB<br />Runtime: 1hr 30 min<br /><br />From the three-time Academy Award®-nominated producer of The Wolf of Wall Street, The Good Mother follows journalist Marissa Bennings who, after the murder of her estranged son, forms an unlikely alliance with his pregnant girlfriend Paige to track down the killers. Together they confront a world of corruption and drugs in the seedy underbelly of a small city in upstate New York. As they get closer to the truth, they unearth an even darker secret. Starring two-time Academy Award®-winner Hilary Swank, Olivia Cooke and Jack Reynor.',
-      link: 'https://yts.mx/movies/the-good-mother-2023',
-      published: 1695171147000,
-      created: 1695171147000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/3103030474806062054EFFF653D6B4B4AC8E92E6',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'The Good Mother (2023) [1080p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/the-good-mother-2023"><img src="https://img.yts.mx/assets/images/movies/the_good_mother_2023/medium-cover.jpg" alt="The Good Mother (2023)" /></a><br />IMDB Rating: 4.7/10<br />Genre: Action / Crime / Drama / Mystery / Thriller<br />Size: 1.67 GB<br />Runtime: 1hr 30 min<br /><br />From the three-time Academy Award®-nominated producer of The Wolf of Wall Street, The Good Mother follows journalist Marissa Bennings who, after the murder of her estranged son, forms an unlikely alliance with his pregnant girlfriend Paige to track down the killers. Together they confront a world of corruption and drugs in the seedy underbelly of a small city in upstate New York. As they get closer to the truth, they unearth an even darker secret. Starring two-time Academy Award®-winner Hilary Swank, Olivia Cooke and Jack Reynor.',
-      link: 'https://yts.mx/movies/the-good-mother-2023',
-      published: 1695174455000,
-      created: 1695174455000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/1E7F2A2403FE7A1E5ADDE7046DA0A33357F323D4',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'Wake of the Red Witch (1948) [720p] [BluRay] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/wake-of-the-red-witch-1948"><img src="https://img.yts.mx/assets/images/movies/wake_of_the_red_witch_1948/medium-cover.jpg" alt="Wake of the Red Witch (1948)" /></a><br />IMDB Rating: 6.5/10<br />Genre: Action / Adventure / Romance<br />Size: 983.14 MB<br />Runtime: 1hr 46 min<br /><br />Captain Ralls fights Dutch shipping magnate Mayrant Sidneye for the woman he loves, Angelique Desaix, and for a fortune in gold aboard the Red Witch.—Jim Beaver',
-      link: 'https://yts.mx/movies/wake-of-the-red-witch-1948',
-      published: 1708098913000,
-      created: 1708098913000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/BBC75D3464EA5D131211ABADCDA5073551823F53',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'Wake of the Red Witch (1948) [1080p] [BluRay] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/wake-of-the-red-witch-1948"><img src="https://img.yts.mx/assets/images/movies/wake_of_the_red_witch_1948/medium-cover.jpg" alt="Wake of the Red Witch (1948)" /></a><br />IMDB Rating: 6.5/10<br />Genre: Action / Adventure / Romance<br />Size: 1.78 GB<br />Runtime: 1hr 46 min<br /><br />Captain Ralls fights Dutch shipping magnate Mayrant Sidneye for the woman he loves, Angelique Desaix, and for a fortune in gold aboard the Red Witch.—Jim Beaver',
-      link: 'https://yts.mx/movies/wake-of-the-red-witch-1948',
-      published: 1708102663000,
-      created: 1708102663000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/0639874568B8E20941E5AF8803EBDC7521F1667E',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title:
-        "Bobi Wine: The People's President (2022) [720p] [WEBRip] [YTS.MX]",
-      description:
-        '<a href="https://yts.mx/movies/bobi-wine-the-peoples-president-2022"><img src="https://img.yts.mx/assets/images/movies/bobi_wine_the_peoples_president_2022/medium-cover.jpg" alt="Bobi Wine: The People\'s President (2022)" /></a><br />IMDB Rating: 7.2/10<br />Genre: Documentary<br />Size: 1.02 GB<br />Runtime: 1hr 53 min<br /><br />',
-      link: 'https://yts.mx/movies/bobi-wine-the-peoples-president-2022',
-      published: 1708094796000,
-      created: 1708094796000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/4926EF9F4FC5F204E5B9561901F3C299460C6F07',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title:
-        "Bobi Wine: The People's President (2022) [1080p] [WEBRip] [5.1] [YTS.MX]",
-      description:
-        '<a href="https://yts.mx/movies/bobi-wine-the-peoples-president-2022"><img src="https://img.yts.mx/assets/images/movies/bobi_wine_the_peoples_president_2022/medium-cover.jpg" alt="Bobi Wine: The People\'s President (2022)" /></a><br />IMDB Rating: 7.2/10<br />Genre: Documentary<br />Size: 2.1 GB<br />Runtime: 1hr 53 min<br /><br />',
-      link: 'https://yts.mx/movies/bobi-wine-the-peoples-president-2022',
-      published: 1708099617000,
-      created: 1708099617000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/DF3712764C510A5DD6B5F5582C51F3188A05C3C6',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'Bleeding Love (2023) [720p] [WEBRip] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/bleeding-love-2023"><img src="https://img.yts.mx/assets/images/movies/bleeding_love_2023/medium-cover.jpg" alt="Bleeding Love (2023)" /></a><br />IMDB Rating: 6.3/10<br />Genre: Drama<br />Size: 941.44 MB<br />Runtime: 1hr 36 min<br /><br />',
-      link: 'https://yts.mx/movies/bleeding-love-2023',
-      published: 1708091901000,
-      created: 1708091901000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/28CC0D695FF4A8E677114EF0A9C9BD02FE0FAAB2',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'Bleeding Love (2023) [1080p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/bleeding-love-2023"><img src="https://img.yts.mx/assets/images/movies/bleeding_love_2023/medium-cover.jpg" alt="Bleeding Love (2023)" /></a><br />IMDB Rating: 6.3/10<br />Genre: Drama<br />Size: 1.89 GB<br />Runtime: 1hr 36 min<br /><br />',
-      link: 'https://yts.mx/movies/bleeding-love-2023',
-      published: 1708096500000,
-      created: 1708096500000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/139191CE289751C26AD9DC47E3585885E58954BF',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'Midnight Peepshow (2022) [720p] [WEBRip] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/midnight-peepshow-2022"><img src="https://img.yts.mx/assets/images/movies/midnight_peepshow_2022/medium-cover.jpg" alt="Midnight Peepshow (2022)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Horror<br />Size: 893.74 MB<br />Runtime: 1hr 33 min<br /><br />',
-      link: 'https://yts.mx/movies/midnight-peepshow-2022',
-      published: 1708088733000,
-      created: 1708088733000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/CFAC71092E799BD88767C8DA9BD62FC053B7AA21',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'Midnight Peepshow (2022) [1080p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/midnight-peepshow-2022"><img src="https://img.yts.mx/assets/images/movies/midnight_peepshow_2022/medium-cover.jpg" alt="Midnight Peepshow (2022)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Horror<br />Size: 1.79 GB<br />Runtime: 1hr 33 min<br /><br />',
-      link: 'https://yts.mx/movies/midnight-peepshow-2022',
-      published: 1708091964000,
-      created: 1708091964000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/DDEB015CE98115D623C62AE0E373DF7803FAB15E',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'No Way Up (2024) [720p] [WEBRip] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/no-way-up-2024"><img src="https://img.yts.mx/assets/images/movies/no_way_up_2024/medium-cover.jpg" alt="No Way Up (2024)" /></a><br />IMDB Rating: 4.9/10<br />Genre: Action / Adventure / Drama / Thriller<br />Size: 829.45 MB<br />Runtime: 1hr 30 min<br /><br />Characters from very different backgrounds are thrown together when the plane they\'re travelling on crashes into the Pacific Ocean. When the airliner comes to rest perilously close to the edge of a ravine with the surviving passengers and crew trapped in an air pocket, a nightmare fight for survival ensues with the air supply running out and dangers creeping in from all sides.—Altitude Film Entertainment',
-      link: 'https://yts.mx/movies/no-way-up-2024',
-      published: 1708085983000,
-      created: 1708085983000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/0D075E5BE1D912AAFB32DC152927231ED70771EF',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'No Way Up (2024) [1080p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/no-way-up-2024"><img src="https://img.yts.mx/assets/images/movies/no_way_up_2024/medium-cover.jpg" alt="No Way Up (2024)" /></a><br />IMDB Rating: 4.9/10<br />Genre: Action / Adventure / Drama / Thriller<br />Size: 1.67 GB<br />Runtime: 1hr 30 min<br /><br />Characters from very different backgrounds are thrown together when the plane they\'re travelling on crashes into the Pacific Ocean. When the airliner comes to rest perilously close to the edge of a ravine with the surviving passengers and crew trapped in an air pocket, a nightmare fight for survival ensues with the air supply running out and dangers creeping in from all sides.—Altitude Film Entertainment',
-      link: 'https://yts.mx/movies/no-way-up-2024',
-      published: 1708089560000,
-      created: 1708089560000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/EC61301DE1A9CBCC434048BDC29DAAF1D14C7C84',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-    {
-      title: 'No Way Up (2024) [2160p] [WEBRip] [5.1] [YTS.MX]',
-      description:
-        '<a href="https://yts.mx/movies/no-way-up-2024"><img src="https://img.yts.mx/assets/images/movies/no_way_up_2024/medium-cover.jpg" alt="No Way Up (2024)" /></a><br />IMDB Rating: 4.9/10<br />Genre: Action / Adventure / Drama / Thriller<br />Size: 1.67 GB<br />Runtime: 1hr 30 min<br /><br />Characters from very different backgrounds are thrown together when the plane they\'re travelling on crashes into the Pacific Ocean. When the airliner comes to rest perilously close to the edge of a ravine with the surviving passengers and crew trapped in an air pocket, a nightmare fight for survival ensues with the air supply running out and dangers creeping in from all sides.—Altitude Film Entertainment',
-      link: 'https://yts.mx/movies/no-way-up-2024',
-      published: 1708089560000,
-      created: 1708089560000,
-      category: [],
-      enclosures: [
-        {
-          url: 'https://yts.mx/torrent/download/EC61301DE1A9CBCC434048BDC29DAAF1D14C7C84',
-          type: 'application/x-bittorrent',
-          length: '10000',
-        },
-      ],
-      media: {},
-    },
-  ],
-};
-
-export const xlmJSON = {
-  '?xml': { version: '1.0', encoding: 'UTF-8' },
-  rss: {
-    channel: {
-      'atom:link': {
-        href: 'https://yts.mx/rss',
-        rel: 'self',
-        type: 'application/rss+xml',
-      },
-      title: 'YTS RSS',
-      description: 'Most popular Torrents in the smallest file size RSS Feed',
-      link: 'https://yts.mx/',
-      language: 'en-us',
-      lastBuildDate: 'Sun, 18 Feb 2024 18:52:50 +0100',
-      item: [
-        {
-          title: 'The Abyss (2023) [720p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/the-abyss-2023"><img src="https://img.yts.mx/assets/images/movies/the_abyss_2023/medium-cover.jpg" alt="The Abyss (2023)" /></a><br />IMDB Rating: 5.6/10<br />Genre: Action / Drama / Thriller<br />Size: 947.37 MB<br />Runtime: 1hr 43 min<br /><br />',
-          link: 'https://yts.mx/movies/the-abyss-2023',
-          guid: 'https://yts.mx/movies/the-abyss-2023#720p',
-          pubDate: 'Sun, 18 Feb 2024 18:32:52 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/B481B1BCAAE6EDBBD801A53DCB59F1DC686F11DA',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Bodyguard Seduction (2022) [720p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/bodyguard-seduction-2022"><img src="https://img.yts.mx/assets/images/movies/bodyguard_seduction_2022/medium-cover.jpg" alt="Bodyguard Seduction (2022)" /></a><br />IMDB Rating: 4.7/10<br />Genre: Mystery / Thriller<br />Size: 788.4 MB<br />Runtime: 1hr 30 min<br /><br />',
-          link: 'https://yts.mx/movies/bodyguard-seduction-2022',
-          guid: 'https://yts.mx/movies/bodyguard-seduction-2022#720p',
-          pubDate: 'Sun, 18 Feb 2024 16:58:08 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/1BBDFE20F156A88235180DA4600B4BDEFAEA2CEA',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Bodyguard Seduction (2022) [1080p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/bodyguard-seduction-2022"><img src="https://img.yts.mx/assets/images/movies/bodyguard_seduction_2022/medium-cover.jpg" alt="Bodyguard Seduction (2022)" /></a><br />IMDB Rating: 4.7/10<br />Genre: Mystery / Thriller<br />Size: 1.43 GB<br />Runtime: 1hr 30 min<br /><br />',
-          link: 'https://yts.mx/movies/bodyguard-seduction-2022',
-          guid: 'https://yts.mx/movies/bodyguard-seduction-2022#1080p',
-          pubDate: 'Sun, 18 Feb 2024 17:55:01 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/461B4BD322B768DA26141F2F27DA533A6AFA7D75',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Calamity of Snakes (1982) [720p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/calamity-of-snakes-1982"><img src="https://img.yts.mx/assets/images/movies/calamity_of_snakes_1982/medium-cover.jpg" alt="Calamity of Snakes (1982)" /></a><br />IMDB Rating: 5.6/10<br />Genre: Action / Comedy / Horror<br />Size: 793.05 MB<br />Runtime: 1hr 26 min<br /><br />After a businessman kills thousands of snakes in a pit where they are building a new apartment building, the residents of the building are attacked by thousands of snakes as vengeance.',
-          link: 'https://yts.mx/movies/calamity-of-snakes-1982',
-          guid: 'https://yts.mx/movies/calamity-of-snakes-1982#720p',
-          pubDate: 'Sun, 18 Feb 2024 15:11:09 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/264147B2611177FF0B32D25234940B3F2A943414',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Calamity of Snakes (1982) [1080p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/calamity-of-snakes-1982"><img src="https://img.yts.mx/assets/images/movies/calamity_of_snakes_1982/medium-cover.jpg" alt="Calamity of Snakes (1982)" /></a><br />IMDB Rating: 5.6/10<br />Genre: Action / Comedy / Horror<br />Size: 1.44 GB<br />Runtime: 1hr 26 min<br /><br />After a businessman kills thousands of snakes in a pit where they are building a new apartment building, the residents of the building are attacked by thousands of snakes as vengeance.',
-          link: 'https://yts.mx/movies/calamity-of-snakes-1982',
-          guid: 'https://yts.mx/movies/calamity-of-snakes-1982#1080p',
-          pubDate: 'Sun, 18 Feb 2024 16:12:53 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/FBFFC8A90AF8A8B6502A4DD7980421247DC61AF2',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Calamity of Snakes (1982) [720p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/calamity-of-snakes-1982"><img src="https://img.yts.mx/assets/images/movies/calamity_of_snakes_1982/medium-cover.jpg" alt="Calamity of Snakes (1982)" /></a><br />IMDB Rating: 5.6/10<br />Genre: Action / Comedy / Horror<br />Size: 792.37 MB<br />Runtime: 1hr 26 min<br /><br />After a businessman kills thousands of snakes in a pit where they are building a new apartment building, the residents of the building are attacked by thousands of snakes as vengeance.',
-          link: 'https://yts.mx/movies/calamity-of-snakes-1982',
-          guid: 'https://yts.mx/movies/calamity-of-snakes-1982#720p',
-          pubDate: 'Mon, 08 May 2023 19:38:26 +0200',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/3C2D3BCA9DDD94D371ABE6A9D15991D4D0AA6999',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Calamity of Snakes (1982) [1080p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/calamity-of-snakes-1982"><img src="https://img.yts.mx/assets/images/movies/calamity_of_snakes_1982/medium-cover.jpg" alt="Calamity of Snakes (1982)" /></a><br />IMDB Rating: 5.6/10<br />Genre: Action / Comedy / Horror<br />Size: 1.44 GB<br />Runtime: 1hr 26 min<br /><br />After a businessman kills thousands of snakes in a pit where they are building a new apartment building, the residents of the building are attacked by thousands of snakes as vengeance.',
-          link: 'https://yts.mx/movies/calamity-of-snakes-1982',
-          guid: 'https://yts.mx/movies/calamity-of-snakes-1982#1080p',
-          pubDate: 'Mon, 08 May 2023 21:49:30 +0200',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/73CB50AF498F7885B4CDC15EC6FA5E53AAA81917',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Anyone But You (2023) [720p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/anyone-but-you-2023"><img src="https://img.yts.mx/assets/images/movies/anyone_but_you_2023/medium-cover.jpg" alt="Anyone But You (2023)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Comedy / Romance<br />Size: 951.22 MB<br />Runtime: 1hr 43 min<br /><br />In the aftermath of a captivating first encounter, Bea and Ben discover that their initial fiery passion has inexplicably dwindled to frosty misunderstanding. However, fate intervenes: Thrown together again at a dreamy Australian wedding, they opt for a charade of coupledom. But sparks reignite amid the sun-kissed scenery, forcing them to confront their true feelings and embrace a second chance at love.—Raihan-Khondker',
-          link: 'https://yts.mx/movies/anyone-but-you-2023',
-          guid: 'https://yts.mx/movies/anyone-but-you-2023#720p',
-          pubDate: 'Sun, 18 Feb 2024 09:35:54 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/478AA2D5DC68C6AE4B276470C30CE6F328DAB53E',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Anyone But You (2023) [1080p] [WEBRip] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/anyone-but-you-2023"><img src="https://img.yts.mx/assets/images/movies/anyone_but_you_2023/medium-cover.jpg" alt="Anyone But You (2023)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Comedy / Romance<br />Size: 1.91 GB<br />Runtime: 1hr 43 min<br /><br />In the aftermath of a captivating first encounter, Bea and Ben discover that their initial fiery passion has inexplicably dwindled to frosty misunderstanding. However, fate intervenes: Thrown together again at a dreamy Australian wedding, they opt for a charade of coupledom. But sparks reignite amid the sun-kissed scenery, forcing them to confront their true feelings and embrace a second chance at love.—Raihan-Khondker',
-          link: 'https://yts.mx/movies/anyone-but-you-2023',
-          guid: 'https://yts.mx/movies/anyone-but-you-2023#1080p',
-          pubDate: 'Sun, 18 Feb 2024 10:36:17 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/DF70A15A89CEDE0957A0C4DF685DF08A4F9D680D',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title:
-            'Anyone But You (2023) [1080p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/anyone-but-you-2023"><img src="https://img.yts.mx/assets/images/movies/anyone_but_you_2023/medium-cover.jpg" alt="Anyone But You (2023)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Comedy / Romance<br />Size: 1.73 GB<br />Runtime: 1hr 43 min<br /><br />In the aftermath of a captivating first encounter, Bea and Ben discover that their initial fiery passion has inexplicably dwindled to frosty misunderstanding. However, fate intervenes: Thrown together again at a dreamy Australian wedding, they opt for a charade of coupledom. But sparks reignite amid the sun-kissed scenery, forcing them to confront their true feelings and embrace a second chance at love.—Raihan-Khondker',
-          link: 'https://yts.mx/movies/anyone-but-you-2023',
-          guid: 'https://yts.mx/movies/anyone-but-you-2023#1080p',
-          pubDate: 'Sun, 18 Feb 2024 14:40:08 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/CC089FF26DE6AC5E99ECE4F716F58892BA8EBF89',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Cause for Alarm! (1951) [720p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/cause-for-alarm-1951"><img src="https://img.yts.mx/assets/images/movies/cause_for_alarm_1951/medium-cover.jpg" alt="Cause for Alarm! (1951)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Crime / Drama / Film-Noir / Thriller<br />Size: 686.23 MB<br />Runtime: 1hr 14 min<br /><br />Invalid George Jones (Barry Sullivan) is both physically and mentally ill. He mistakenly believes his wife Ellen (Loretta Young) and his doctor (Bruce Cowling) are having an affair and also planning to kill him. He writes a letter to the district attorney detailing their alleged murder plot. After he has Ellen give the letter to their postman, he reveals its contents to her and then threatens her with a gun. The excitement proves to much and George suffers a fatal collapse. Now Ellen must find a way to retrieve the incriminating letter.—Daniel Bubbeo',
-          link: 'https://yts.mx/movies/cause-for-alarm-1951',
-          guid: 'https://yts.mx/movies/cause-for-alarm-1951#720p',
-          pubDate: 'Sun, 18 Feb 2024 13:48:53 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/CB7E1F4733948A50D974A890950EACF03395ED35',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Cause for Alarm! (1951) [1080p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/cause-for-alarm-1951"><img src="https://img.yts.mx/assets/images/movies/cause_for_alarm_1951/medium-cover.jpg" alt="Cause for Alarm! (1951)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Crime / Drama / Film-Noir / Thriller<br />Size: 1.24 GB<br />Runtime: 1hr 14 min<br /><br />Invalid George Jones (Barry Sullivan) is both physically and mentally ill. He mistakenly believes his wife Ellen (Loretta Young) and his doctor (Bruce Cowling) are having an affair and also planning to kill him. He writes a letter to the district attorney detailing their alleged murder plot. After he has Ellen give the letter to their postman, he reveals its contents to her and then threatens her with a gun. The excitement proves to much and George suffers a fatal collapse. Now Ellen must find a way to retrieve the incriminating letter.—Daniel Bubbeo',
-          link: 'https://yts.mx/movies/cause-for-alarm-1951',
-          guid: 'https://yts.mx/movies/cause-for-alarm-1951#1080p',
-          pubDate: 'Sun, 18 Feb 2024 14:36:31 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/4163A72C080E7C00B8AC175FBA795F684EEA261E',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Tremors (1990) [720p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/tremors-1990"><img src="https://img.yts.mx/assets/images/movies/Tremors_1990/medium-cover.jpg" alt="Tremors (1990)" /></a><br />IMDB Rating: 7.1/10<br />Genre: Action / Comedy / Horror / Sci-Fi<br />Size: 699.56 MB<br />Runtime: 1hr 36 min<br /><br />Trying to escape their dead-end life in the desert town of Perfection, Nevada, Valentine "Val" McKee and Earl Bassett find themselves sidetracked when corpses mysteriously start piling up around them. When Val and Earl have a run-in with some creepy tentacled creatures that have made lunch out of a road crew, they retreat back to the town to spread the news. Although these monsters are not exactly intellectuals, their mental powers are still mind-boggling. Detecting a human hiding in a car, they dig under the vehicle, causing it to sink into the earth. Val and Earl get help from several townspeople. With the help of Rhonda, Val and Earl kill one of the monsters. But there are three of them left, and they are each about 30 feet long. Walter starts calling the monsters graboids. When the survivors think they have outwitted the graboids by taking refuge atop the roofs of buildings, the graboids merely destroy the buildings\' foundations, killing two people. The graboids are quick learners when it comes to stalking their prey, and the humans must continually be on their toes if they are to stay alive while they try to find a way to defeat the graboids.—Todd Baldridge',
-          link: 'https://yts.mx/movies/tremors-1990',
-          guid: 'https://yts.mx/movies/tremors-1990#720p',
-          pubDate: 'Sun, 01 Nov 2015 03:47:07 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/5372C63146C6BDD8950E4D47FAC6703C10714939',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Tremors (1990) [1080p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/tremors-1990"><img src="https://img.yts.mx/assets/images/movies/Tremors_1990/medium-cover.jpg" alt="Tremors (1990)" /></a><br />IMDB Rating: 7.1/10<br />Genre: Action / Comedy / Horror / Sci-Fi<br />Size: 1.30 GB<br />Runtime: 1hr 36 min<br /><br />Trying to escape their dead-end life in the desert town of Perfection, Nevada, Valentine "Val" McKee and Earl Bassett find themselves sidetracked when corpses mysteriously start piling up around them. When Val and Earl have a run-in with some creepy tentacled creatures that have made lunch out of a road crew, they retreat back to the town to spread the news. Although these monsters are not exactly intellectuals, their mental powers are still mind-boggling. Detecting a human hiding in a car, they dig under the vehicle, causing it to sink into the earth. Val and Earl get help from several townspeople. With the help of Rhonda, Val and Earl kill one of the monsters. But there are three of them left, and they are each about 30 feet long. Walter starts calling the monsters graboids. When the survivors think they have outwitted the graboids by taking refuge atop the roofs of buildings, the graboids merely destroy the buildings\' foundations, killing two people. The graboids are quick learners when it comes to stalking their prey, and the humans must continually be on their toes if they are to stay alive while they try to find a way to defeat the graboids.—Todd Baldridge',
-          link: 'https://yts.mx/movies/tremors-1990',
-          guid: 'https://yts.mx/movies/tremors-1990#1080p',
-          pubDate: 'Sun, 01 Nov 2015 03:47:09 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/4FC917FFD092801AFF50FE1E568ED7D0504E336C',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title:
-            'Tremors (1990) [2160p] [BluRay] [x265] [10bit] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/tremors-1990"><img src="https://img.yts.mx/assets/images/movies/Tremors_1990/medium-cover.jpg" alt="Tremors (1990)" /></a><br />IMDB Rating: 7.1/10<br />Genre: Action / Comedy / Horror / Sci-Fi<br />Size: 4.28 GB<br />Runtime: 1hr 36 min<br /><br />Trying to escape their dead-end life in the desert town of Perfection, Nevada, Valentine "Val" McKee and Earl Bassett find themselves sidetracked when corpses mysteriously start piling up around them. When Val and Earl have a run-in with some creepy tentacled creatures that have made lunch out of a road crew, they retreat back to the town to spread the news. Although these monsters are not exactly intellectuals, their mental powers are still mind-boggling. Detecting a human hiding in a car, they dig under the vehicle, causing it to sink into the earth. Val and Earl get help from several townspeople. With the help of Rhonda, Val and Earl kill one of the monsters. But there are three of them left, and they are each about 30 feet long. Walter starts calling the monsters graboids. When the survivors think they have outwitted the graboids by taking refuge atop the roofs of buildings, the graboids merely destroy the buildings\' foundations, killing two people. The graboids are quick learners when it comes to stalking their prey, and the humans must continually be on their toes if they are to stay alive while they try to find a way to defeat the graboids.—Todd Baldridge',
-          link: 'https://yts.mx/movies/tremors-1990',
-          guid: 'https://yts.mx/movies/tremors-1990#2160p',
-          pubDate: 'Sun, 18 Feb 2024 14:00:27 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/7E05139707F19A0222B5C5EB0F89BD3B18FACFBB',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'The Sidemen Story (2024) [720p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/the-sidemen-story-2024"><img src="https://img.yts.mx/assets/images/movies/the_sidemen_story_2024/medium-cover.jpg" alt="The Sidemen Story (2024)" /></a><br />IMDB Rating: 6.9/10<br />Genre: Biography / Documentary<br />Size: 917.58 MB<br />Runtime: 1hr 39 min<br /><br />',
-          link: 'https://yts.mx/movies/the-sidemen-story-2024',
-          guid: 'https://yts.mx/movies/the-sidemen-story-2024#720p',
-          pubDate: 'Sun, 18 Feb 2024 09:36:57 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/7FC82D356B278C41D345D1C77B28061FE7256F69',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'The Sidemen Story (2024) [1080p] [WEBRip] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/the-sidemen-story-2024"><img src="https://img.yts.mx/assets/images/movies/the_sidemen_story_2024/medium-cover.jpg" alt="The Sidemen Story (2024)" /></a><br />IMDB Rating: 6.9/10<br />Genre: Biography / Documentary<br />Size: 1.84 GB<br />Runtime: 1hr 39 min<br /><br />',
-          link: 'https://yts.mx/movies/the-sidemen-story-2024',
-          guid: 'https://yts.mx/movies/the-sidemen-story-2024#1080p',
-          pubDate: 'Sun, 18 Feb 2024 10:58:40 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/37BC252B31034ECCCC8B2F5AB1454CF54FEB33B6',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Nothing But a Man (1964) [720p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/nothing-but-a-man-1964"><img src="https://img.yts.mx/assets/images/movies/nothing_but_a_man_1964/medium-cover.jpg" alt="Nothing But a Man (1964)" /></a><br />IMDB Rating: 7.9/10<br />Genre: Action / Drama / Romance<br />Size: 841.39 MB<br />Runtime: 1hr 35 min<br /><br />Born in Birmingham, Duff Anderson, the father of a male toddler, who lives with a nanny, re-locates to a small town to work on the railroad. He meets with and is attracted to Josie much to the chagrin of her preacher father. The marriage does take place nevertheless, both re-locate to live in their own house and he gets a job in a mill. He decides not to bring his son to live with them. Challenges arise when the Mill Foreman finds out that Duff is attempting to unionize the workers, forcing Duff to quit, and look for work elsewhere. Unable to reconcile himself to working on a daily wage of $2.50 picking cotton nor even as a waiter, he gets a job at a garage. He is enraged at a customer for belittling him and Josie, and is let go. Unemployed, unable to support his wife and son, he gets abusive and leaves - perhaps never to return.—rAjOo (gunwanti@hotmail.com)',
-          link: 'https://yts.mx/movies/nothing-but-a-man-1964',
-          guid: 'https://yts.mx/movies/nothing-but-a-man-1964#720p',
-          pubDate: 'Sun, 18 Feb 2024 09:47:56 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/4DBA9B04093E08A319BC368E2270489560EFEBA0',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Nothing But a Man (1964) [1080p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/nothing-but-a-man-1964"><img src="https://img.yts.mx/assets/images/movies/nothing_but_a_man_1964/medium-cover.jpg" alt="Nothing But a Man (1964)" /></a><br />IMDB Rating: 7.9/10<br />Genre: Action / Drama / Romance<br />Size: 1.53 GB<br />Runtime: 1hr 35 min<br /><br />Born in Birmingham, Duff Anderson, the father of a male toddler, who lives with a nanny, re-locates to a small town to work on the railroad. He meets with and is attracted to Josie much to the chagrin of her preacher father. The marriage does take place nevertheless, both re-locate to live in their own house and he gets a job in a mill. He decides not to bring his son to live with them. Challenges arise when the Mill Foreman finds out that Duff is attempting to unionize the workers, forcing Duff to quit, and look for work elsewhere. Unable to reconcile himself to working on a daily wage of $2.50 picking cotton nor even as a waiter, he gets a job at a garage. He is enraged at a customer for belittling him and Josie, and is let go. Unemployed, unable to support his wife and son, he gets abusive and leaves - perhaps never to return.—rAjOo (gunwanti@hotmail.com)',
-          link: 'https://yts.mx/movies/nothing-but-a-man-1964',
-          guid: 'https://yts.mx/movies/nothing-but-a-man-1964#1080p',
-          pubDate: 'Sun, 18 Feb 2024 10:50:53 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/B49A191BF5E983B49BD7CFF6C73514F3964A5033',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Nothing But a Man (1964) [720p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/nothing-but-a-man-1964"><img src="https://img.yts.mx/assets/images/movies/nothing_but_a_man_1964/medium-cover.jpg" alt="Nothing But a Man (1964)" /></a><br />IMDB Rating: 7.9/10<br />Genre: Action / Drama / Romance<br />Size: 845.53 MB<br />Runtime: 1hr 35 min<br /><br />Born in Birmingham, Duff Anderson, the father of a male toddler, who lives with a nanny, re-locates to a small town to work on the railroad. He meets with and is attracted to Josie much to the chagrin of her preacher father. The marriage does take place nevertheless, both re-locate to live in their own house and he gets a job in a mill. He decides not to bring his son to live with them. Challenges arise when the Mill Foreman finds out that Duff is attempting to unionize the workers, forcing Duff to quit, and look for work elsewhere. Unable to reconcile himself to working on a daily wage of $2.50 picking cotton nor even as a waiter, he gets a job at a garage. He is enraged at a customer for belittling him and Josie, and is let go. Unemployed, unable to support his wife and son, he gets abusive and leaves - perhaps never to return.—rAjOo (gunwanti@hotmail.com)',
-          link: 'https://yts.mx/movies/nothing-but-a-man-1964',
-          guid: 'https://yts.mx/movies/nothing-but-a-man-1964#720p',
-          pubDate: 'Fri, 27 May 2022 21:10:12 +0200',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/DA9CCA19B44249B542BB6F6C973FD14EB82E7133',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Nothing But a Man (1964) [1080p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/nothing-but-a-man-1964"><img src="https://img.yts.mx/assets/images/movies/nothing_but_a_man_1964/medium-cover.jpg" alt="Nothing But a Man (1964)" /></a><br />IMDB Rating: 7.9/10<br />Genre: Action / Drama / Romance<br />Size: 1.53 GB<br />Runtime: 1hr 35 min<br /><br />Born in Birmingham, Duff Anderson, the father of a male toddler, who lives with a nanny, re-locates to a small town to work on the railroad. He meets with and is attracted to Josie much to the chagrin of her preacher father. The marriage does take place nevertheless, both re-locate to live in their own house and he gets a job in a mill. He decides not to bring his son to live with them. Challenges arise when the Mill Foreman finds out that Duff is attempting to unionize the workers, forcing Duff to quit, and look for work elsewhere. Unable to reconcile himself to working on a daily wage of $2.50 picking cotton nor even as a waiter, he gets a job at a garage. He is enraged at a customer for belittling him and Josie, and is let go. Unemployed, unable to support his wife and son, he gets abusive and leaves - perhaps never to return.—rAjOo (gunwanti@hotmail.com)',
-          link: 'https://yts.mx/movies/nothing-but-a-man-1964',
-          guid: 'https://yts.mx/movies/nothing-but-a-man-1964#1080p',
-          pubDate: 'Fri, 27 May 2022 22:44:36 +0200',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/BA6BB2AD15868C080C98064FBA745FC69ABDD7CC',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Story Ave (2023) [720p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/story-ave-2023"><img src="https://img.yts.mx/assets/images/movies/story_ave_2023/medium-cover.jpg" alt="Story Ave (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama<br />Size: 979.08 MB<br />Runtime: 1hr 46 min<br /><br />',
-          link: 'https://yts.mx/movies/story-ave-2023',
-          guid: 'https://yts.mx/movies/story-ave-2023#720p',
-          pubDate: 'Sun, 18 Feb 2024 07:48:59 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/FEBB30242157C8EA08B09B98FE3EFF98486ACFEF',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Story Ave (2023) [1080p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/story-ave-2023"><img src="https://img.yts.mx/assets/images/movies/story_ave_2023/medium-cover.jpg" alt="Story Ave (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama<br />Size: 1.77 GB<br />Runtime: 1hr 46 min<br /><br />',
-          link: 'https://yts.mx/movies/story-ave-2023',
-          guid: 'https://yts.mx/movies/story-ave-2023#1080p',
-          pubDate: 'Sun, 18 Feb 2024 09:02:35 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/BCC54D6FE8AE4DB22F87B3743E9D12DC44086F4F',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Story Ave (2023) [720p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/story-ave-2023"><img src="https://img.yts.mx/assets/images/movies/story_ave_2023/medium-cover.jpg" alt="Story Ave (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama<br />Size: 979.48 MB<br />Runtime: 1hr 46 min<br /><br />',
-          link: 'https://yts.mx/movies/story-ave-2023',
-          guid: 'https://yts.mx/movies/story-ave-2023#720p',
-          pubDate: 'Tue, 06 Feb 2024 08:55:20 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/D46BCC209A3617639AB631DC5EA92E8A12FB368E',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Story Ave (2023) [1080p] [WEBRip] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/story-ave-2023"><img src="https://img.yts.mx/assets/images/movies/story_ave_2023/medium-cover.jpg" alt="Story Ave (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama<br />Size: 1.97 GB<br />Runtime: 1hr 46 min<br /><br />',
-          link: 'https://yts.mx/movies/story-ave-2023',
-          guid: 'https://yts.mx/movies/story-ave-2023#1080p',
-          pubDate: 'Tue, 06 Feb 2024 10:00:46 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/618E90EE3C619B74053007A47A6690CAF0A7654E',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Eileen (2023) [720p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 895.96 MB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
-          link: 'https://yts.mx/movies/eileen-2023',
-          guid: 'https://yts.mx/movies/eileen-2023#720p',
-          pubDate: 'Sun, 18 Feb 2024 07:45:09 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/172D82D76C74087363E5D164D921E2ACCE0C979A',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Eileen (2023) [1080p] [BluRay] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 1.8 GB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
-          link: 'https://yts.mx/movies/eileen-2023',
-          guid: 'https://yts.mx/movies/eileen-2023#1080p',
-          pubDate: 'Sun, 18 Feb 2024 09:01:01 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/49C22DB047A202F87C8AF313452F9A12466F79B4',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Eileen (2023) [720p] [WEBRip] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 896.54 MB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
-          link: 'https://yts.mx/movies/eileen-2023',
-          guid: 'https://yts.mx/movies/eileen-2023#720p',
-          pubDate: 'Tue, 02 Jan 2024 07:04:08 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/0F91C0492E41E08CBB76810D484FB105DBF91D8A',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Eileen (2023) [1080p] [WEBRip] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 1.8 GB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
-          link: 'https://yts.mx/movies/eileen-2023',
-          guid: 'https://yts.mx/movies/eileen-2023#1080p',
-          pubDate: 'Tue, 02 Jan 2024 08:15:25 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/D7AE2DC467B381F5C8A86FF91F1846B722C8E3E0',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Eileen (2023) [1080p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 1.63 GB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
-          link: 'https://yts.mx/movies/eileen-2023',
-          guid: 'https://yts.mx/movies/eileen-2023#1080p',
-          pubDate: 'Tue, 02 Jan 2024 13:47:24 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/0E7F3409C048AFA061915905556B1E6F2F85ACE8',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Eileen (2023) [2160p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 4.34 GB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
-          link: 'https://yts.mx/movies/eileen-2023',
-          guid: 'https://yts.mx/movies/eileen-2023#2160p',
-          pubDate: 'Thu, 04 Jan 2024 00:28:04 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/CA6121F4BF489BC13E57152EE1A3187360D2E57C',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'The Selling (2011) [720p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/the-selling-2011"><img src="https://img.yts.mx/assets/images/movies/the_selling_2011/medium-cover.jpg" alt="The Selling (2011)" /></a><br />IMDB Rating: 6.0/10<br />Genre: Comedy<br />Size: 826.11 MB<br />Runtime: 1hr 32 min<br /><br />',
-          link: 'https://yts.mx/movies/the-selling-2011',
-          guid: 'https://yts.mx/movies/the-selling-2011#720p',
-          pubDate: 'Sun, 18 Feb 2024 05:14:06 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/725A1D7B2BCA4FDD7485021696A672B8ACFE5AD2',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'The Selling (2011) [1080p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/the-selling-2011"><img src="https://img.yts.mx/assets/images/movies/the_selling_2011/medium-cover.jpg" alt="The Selling (2011)" /></a><br />IMDB Rating: 6.0/10<br />Genre: Comedy<br />Size: 1.5 GB<br />Runtime: 1hr 32 min<br /><br />',
-          link: 'https://yts.mx/movies/the-selling-2011',
-          guid: 'https://yts.mx/movies/the-selling-2011#1080p',
-          pubDate: 'Sun, 18 Feb 2024 06:14:27 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/228A9F41684B55E34CECEE2A8151374A35B5FD66',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Repligator (1998) [720p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/repligator-1998"><img src="https://img.yts.mx/assets/images/movies/repligator_1998/medium-cover.jpg" alt="Repligator (1998)" /></a><br />IMDB Rating: 2.6/10<br />Genre: Action / Comedy / Sci-Fi<br />Size: 786.83 MB<br />Runtime: 1hr 26 min<br /><br />',
-          link: 'https://yts.mx/movies/repligator-1998',
-          guid: 'https://yts.mx/movies/repligator-1998#720p',
-          pubDate: 'Sun, 18 Feb 2024 03:39:54 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/C101922D0BD1A7E1CEB544142E780969CFE2687E',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Repligator (1998) [1080p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/repligator-1998"><img src="https://img.yts.mx/assets/images/movies/repligator_1998/medium-cover.jpg" alt="Repligator (1998)" /></a><br />IMDB Rating: 2.6/10<br />Genre: Action / Comedy / Sci-Fi<br />Size: 1.42 GB<br />Runtime: 1hr 26 min<br /><br />',
-          link: 'https://yts.mx/movies/repligator-1998',
-          guid: 'https://yts.mx/movies/repligator-1998#1080p',
-          pubDate: 'Sun, 18 Feb 2024 04:39:20 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/AB7D6B62CC97CA52F53B912FEF25823F43BB53B7',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'This Is Me... Now (2024) [720p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/this-is-me-now-2024"><img src="https://img.yts.mx/assets/images/movies/this_is_me_now_2024/medium-cover.jpg" alt="This Is Me... Now (2024)" /></a><br />IMDB Rating: 5.3/10<br />Genre: Music / Musical<br />Size: 603 MB<br />Runtime: 1hr 5 min<br /><br />',
-          link: 'https://yts.mx/movies/this-is-me-now-2024',
-          guid: 'https://yts.mx/movies/this-is-me-now-2024#720p',
-          pubDate: 'Fri, 16 Feb 2024 01:31:38 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/1026B9FA4E75F11569917E7BF56D7005DAE68CA6',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'This Is Me... Now (2024) [1080p] [WEBRip] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/this-is-me-now-2024"><img src="https://img.yts.mx/assets/images/movies/this_is_me_now_2024/medium-cover.jpg" alt="This Is Me... Now (2024)" /></a><br />IMDB Rating: 5.3/10<br />Genre: Music / Musical<br />Size: 1.21 GB<br />Runtime: 1hr 5 min<br /><br />',
-          link: 'https://yts.mx/movies/this-is-me-now-2024',
-          guid: 'https://yts.mx/movies/this-is-me-now-2024#1080p',
-          pubDate: 'Fri, 16 Feb 2024 02:08:39 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/29452026DB59FB985AC42C9219D066394E1EC04A',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title:
-            'This Is Me... Now (2024) [2160p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/this-is-me-now-2024"><img src="https://img.yts.mx/assets/images/movies/this_is_me_now_2024/medium-cover.jpg" alt="This Is Me... Now (2024)" /></a><br />IMDB Rating: 5.3/10<br />Genre: Music / Musical<br />Size: 2.92 GB<br />Runtime: 1hr 5 min<br /><br />',
-          link: 'https://yts.mx/movies/this-is-me-now-2024',
-          guid: 'https://yts.mx/movies/this-is-me-now-2024#2160p',
-          pubDate: 'Sun, 18 Feb 2024 04:16:08 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/7FE6D31CC25E999D6A09305786D43BA6B7C5F57E',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Has Anybody Seen My Gal (1952) [720p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/has-anybody-seen-my-gal-1952"><img src="https://img.yts.mx/assets/images/movies/has_anybody_seen_my_gal_1952/medium-cover.jpg" alt="Has Anybody Seen My Gal (1952)" /></a><br />IMDB Rating: 7.1/10<br />Genre: Comedy<br />Size: 816.11 MB<br />Runtime: 1hr 28 min<br /><br />Wealthy Samuel Fulton is getting older and has no family of his own. He decides to leave his estate to the family of his first love, who turned down his marriage proposal years ago because he was poor. But he wants to test the family before leaving his money to them. He takes a room in their home and a job in the father\'s shop. He anonymously grants them $100,000. Harriet Blaisdell moves the family into a mansion and makes plans to marry her daughter Millicent off to a socialite rather than her soda jerk boyfriend Dan. The money goes to their heads, and they soon find themselves broke, back in their old house, and back to their old lives. Father back in his shop, Millicent engaged to Dan, and everyone seemingly much happier. Hoping they learned their lesson, Fulton takes his leave of the family.—Ed Stephan',
-          link: 'https://yts.mx/movies/has-anybody-seen-my-gal-1952',
-          guid: 'https://yts.mx/movies/has-anybody-seen-my-gal-1952#720p',
-          pubDate: 'Sat, 17 Feb 2024 22:58:56 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/74FF3CE61FFE874AA862941229BD6DE708B22940',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title: 'Has Anybody Seen My Gal (1952) [1080p] [BluRay] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/has-anybody-seen-my-gal-1952"><img src="https://img.yts.mx/assets/images/movies/has_anybody_seen_my_gal_1952/medium-cover.jpg" alt="Has Anybody Seen My Gal (1952)" /></a><br />IMDB Rating: 7.1/10<br />Genre: Comedy<br />Size: 1.48 GB<br />Runtime: 1hr 28 min<br /><br />Wealthy Samuel Fulton is getting older and has no family of his own. He decides to leave his estate to the family of his first love, who turned down his marriage proposal years ago because he was poor. But he wants to test the family before leaving his money to them. He takes a room in their home and a job in the father\'s shop. He anonymously grants them $100,000. Harriet Blaisdell moves the family into a mansion and makes plans to marry her daughter Millicent off to a socialite rather than her soda jerk boyfriend Dan. The money goes to their heads, and they soon find themselves broke, back in their old house, and back to their old lives. Father back in his shop, Millicent engaged to Dan, and everyone seemingly much happier. Hoping they learned their lesson, Fulton takes his leave of the family.—Ed Stephan',
-          link: 'https://yts.mx/movies/has-anybody-seen-my-gal-1952',
-          guid: 'https://yts.mx/movies/has-anybody-seen-my-gal-1952#1080p',
-          pubDate: 'Sun, 18 Feb 2024 00:00:35 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/854ECFF72F7434E7F165CE6B43BD6B25EB550A5E',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title:
-            'Snoopy Presents: Welcome Home, Franklin (2024) [720p] [WEBRip] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024"><img src="https://img.yts.mx/assets/images/movies/snoopy_presents_welcome_home_franklin_2024/medium-cover.jpg" alt="Snoopy Presents: Welcome Home, Franklin (2024)" /></a><br />IMDB Rating: 7.2/10<br />Genre: Animation<br />Size: 360.67 MB<br />Runtime: 12hr 39 min<br /><br />The origin story for one of Peanuts\' most beloved characters, Franklin, follows how he approaches making new friends. Franklin\'s family is always on the move with his dad\'s military job, and everywhere he goes Franklin finds support in a notebook filled with his grandfather\'s advice on friendship. But when Franklin tries his usual strategies with the Peanuts gang, he has trouble fitting in. That\'s until he learns about the neighborhood Soap Box Derby race - according to his grandfather, everyone loves a winner. He\'s sure that winning the race will also mean winning over some new friends. All he needs is a partner, which he finds in Charlie Brown. Franklin and Charlie Brown work together to build a car and in the process become good buddies. But as the race nears, the pressure mounts - can their car and their newfound friendship make it to the finish line?—Apple TV+',
-          link: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024',
-          guid: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024#720p',
-          pubDate: 'Fri, 16 Feb 2024 03:35:46 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/6857C0A27586A351E9B404756CB77A39B541E51F',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title:
-            'Snoopy Presents: Welcome Home, Franklin (2024) [1080p] [WEBRip] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024"><img src="https://img.yts.mx/assets/images/movies/snoopy_presents_welcome_home_franklin_2024/medium-cover.jpg" alt="Snoopy Presents: Welcome Home, Franklin (2024)" /></a><br />IMDB Rating: 7.2/10<br />Genre: Animation<br />Size: 738.44 MB<br />Runtime: 12hr 39 min<br /><br />The origin story for one of Peanuts\' most beloved characters, Franklin, follows how he approaches making new friends. Franklin\'s family is always on the move with his dad\'s military job, and everywhere he goes Franklin finds support in a notebook filled with his grandfather\'s advice on friendship. But when Franklin tries his usual strategies with the Peanuts gang, he has trouble fitting in. That\'s until he learns about the neighborhood Soap Box Derby race - according to his grandfather, everyone loves a winner. He\'s sure that winning the race will also mean winning over some new friends. All he needs is a partner, which he finds in Charlie Brown. Franklin and Charlie Brown work together to build a car and in the process become good buddies. But as the race nears, the pressure mounts - can their car and their newfound friendship make it to the finish line?—Apple TV+',
-          link: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024',
-          guid: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024#1080p',
-          pubDate: 'Fri, 16 Feb 2024 03:56:42 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/3F5DDC35EF8B304C7AFEC7EF8F8F7C44FAD68EAA',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-        {
-          title:
-            'Snoopy Presents: Welcome Home, Franklin (2024) [2160p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]',
-          description:
-            '<a href="https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024"><img src="https://img.yts.mx/assets/images/movies/snoopy_presents_welcome_home_franklin_2024/medium-cover.jpg" alt="Snoopy Presents: Welcome Home, Franklin (2024)" /></a><br />IMDB Rating: 7.2/10<br />Genre: Animation<br />Size: 1.74 GB<br />Runtime: 12hr 39 min<br /><br />The origin story for one of Peanuts\' most beloved characters, Franklin, follows how he approaches making new friends. Franklin\'s family is always on the move with his dad\'s military job, and everywhere he goes Franklin finds support in a notebook filled with his grandfather\'s advice on friendship. But when Franklin tries his usual strategies with the Peanuts gang, he has trouble fitting in. That\'s until he learns about the neighborhood Soap Box Derby race - according to his grandfather, everyone loves a winner. He\'s sure that winning the race will also mean winning over some new friends. All he needs is a partner, which he finds in Charlie Brown. Franklin and Charlie Brown work together to build a car and in the process become good buddies. But as the race nears, the pressure mounts - can their car and their newfound friendship make it to the finish line?—Apple TV+',
-          link: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024',
-          guid: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024#2160p',
-          pubDate: 'Sat, 17 Feb 2024 21:28:18 +0100',
-          enclosure: {
-            url: 'https://yts.mx/torrent/download/22987BCBE6A58AC7AE393B5326604D047F096273',
-            type: 'application/x-bittorrent',
-            length: '10000',
-          },
-        },
-      ],
-    },
-    version: '2.0',
-    'xmlns:dc': 'http://purl.org/dc/elements/1.1/',
-    'xmlns:content': 'http://purl.org/rss/1.0/modules/content/',
-    'xmlns:atom': 'http://www.w3.org/2005/Atom',
-  },
-};
