+++ test/helpers/list.ts
@@ -1,0 +1,589 @@
+import { XMLYTSResp } from '../../src/types/main';
+
+export const xlmJSON: XMLYTSResp = {
+  '?xml': { version: '1.0', encoding: 'UTF-8' },
+  rss: {
+    channel: {
+      'atom:link': {
+        href: 'https://yts.mx/rss',
+        rel: 'self',
+        type: 'application/rss+xml',
+      },
+      title: 'YTS RSS',
+      description: 'Most popular Torrents in the smallest file size RSS Feed',
+      link: 'https://yts.mx/',
+      language: 'en-us',
+      lastBuildDate: 'Sun, 18 Feb 2024 18:52:50 +0100',
+      item: [
+        {
+          title: 'The Abyss (2023) [720p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/the-abyss-2023"><img src="https://img.yts.mx/assets/images/movies/the_abyss_2023/medium-cover.jpg" alt="The Abyss (2023)" /></a><br />IMDB Rating: 5.6/10<br />Genre: Action / Drama / Thriller<br />Size: 947.37 MB<br />Runtime: 1hr 43 min<br /><br />',
+          link: 'https://yts.mx/movies/the-abyss-2023',
+          guid: 'https://yts.mx/movies/the-abyss-2023#720p',
+          pubDate: 'Sun, 18 Feb 2024 18:32:52 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/B481B1BCAAE6EDBBD801A53DCB59F1DC686F11DA',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Bodyguard Seduction (2022) [720p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/bodyguard-seduction-2022"><img src="https://img.yts.mx/assets/images/movies/bodyguard_seduction_2022/medium-cover.jpg" alt="Bodyguard Seduction (2022)" /></a><br />IMDB Rating: 4.7/10<br />Genre: Mystery / Thriller<br />Size: 788.4 MB<br />Runtime: 1hr 30 min<br /><br />',
+          link: 'https://yts.mx/movies/bodyguard-seduction-2022',
+          guid: 'https://yts.mx/movies/bodyguard-seduction-2022#720p',
+          pubDate: 'Sun, 18 Feb 2024 16:58:08 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/1BBDFE20F156A88235180DA4600B4BDEFAEA2CEA',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Bodyguard Seduction (2022) [1080p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/bodyguard-seduction-2022"><img src="https://img.yts.mx/assets/images/movies/bodyguard_seduction_2022/medium-cover.jpg" alt="Bodyguard Seduction (2022)" /></a><br />IMDB Rating: 4.7/10<br />Genre: Mystery / Thriller<br />Size: 1.43 GB<br />Runtime: 1hr 30 min<br /><br />',
+          link: 'https://yts.mx/movies/bodyguard-seduction-2022',
+          guid: 'https://yts.mx/movies/bodyguard-seduction-2022#1080p',
+          pubDate: 'Sun, 18 Feb 2024 17:55:01 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/461B4BD322B768DA26141F2F27DA533A6AFA7D75',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Calamity of Snakes (1982) [720p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/calamity-of-snakes-1982"><img src="https://img.yts.mx/assets/images/movies/calamity_of_snakes_1982/medium-cover.jpg" alt="Calamity of Snakes (1982)" /></a><br />IMDB Rating: 5.6/10<br />Genre: Action / Comedy / Horror<br />Size: 793.05 MB<br />Runtime: 1hr 26 min<br /><br />After a businessman kills thousands of snakes in a pit where they are building a new apartment building, the residents of the building are attacked by thousands of snakes as vengeance.',
+          link: 'https://yts.mx/movies/calamity-of-snakes-1982',
+          guid: 'https://yts.mx/movies/calamity-of-snakes-1982#720p',
+          pubDate: 'Sun, 18 Feb 2024 15:11:09 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/264147B2611177FF0B32D25234940B3F2A943414',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Calamity of Snakes (1982) [1080p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/calamity-of-snakes-1982"><img src="https://img.yts.mx/assets/images/movies/calamity_of_snakes_1982/medium-cover.jpg" alt="Calamity of Snakes (1982)" /></a><br />IMDB Rating: 5.6/10<br />Genre: Action / Comedy / Horror<br />Size: 1.44 GB<br />Runtime: 1hr 26 min<br /><br />After a businessman kills thousands of snakes in a pit where they are building a new apartment building, the residents of the building are attacked by thousands of snakes as vengeance.',
+          link: 'https://yts.mx/movies/calamity-of-snakes-1982',
+          guid: 'https://yts.mx/movies/calamity-of-snakes-1982#1080p',
+          pubDate: 'Sun, 18 Feb 2024 16:12:53 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/FBFFC8A90AF8A8B6502A4DD7980421247DC61AF2',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Calamity of Snakes (1982) [720p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/calamity-of-snakes-1982"><img src="https://img.yts.mx/assets/images/movies/calamity_of_snakes_1982/medium-cover.jpg" alt="Calamity of Snakes (1982)" /></a><br />IMDB Rating: 5.6/10<br />Genre: Action / Comedy / Horror<br />Size: 792.37 MB<br />Runtime: 1hr 26 min<br /><br />After a businessman kills thousands of snakes in a pit where they are building a new apartment building, the residents of the building are attacked by thousands of snakes as vengeance.',
+          link: 'https://yts.mx/movies/calamity-of-snakes-1982',
+          guid: 'https://yts.mx/movies/calamity-of-snakes-1982#720p',
+          pubDate: 'Mon, 08 May 2023 19:38:26 +0200',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/3C2D3BCA9DDD94D371ABE6A9D15991D4D0AA6999',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Calamity of Snakes (1982) [1080p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/calamity-of-snakes-1982"><img src="https://img.yts.mx/assets/images/movies/calamity_of_snakes_1982/medium-cover.jpg" alt="Calamity of Snakes (1982)" /></a><br />IMDB Rating: 5.6/10<br />Genre: Action / Comedy / Horror<br />Size: 1.44 GB<br />Runtime: 1hr 26 min<br /><br />After a businessman kills thousands of snakes in a pit where they are building a new apartment building, the residents of the building are attacked by thousands of snakes as vengeance.',
+          link: 'https://yts.mx/movies/calamity-of-snakes-1982',
+          guid: 'https://yts.mx/movies/calamity-of-snakes-1982#1080p',
+          pubDate: 'Mon, 08 May 2023 21:49:30 +0200',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/73CB50AF498F7885B4CDC15EC6FA5E53AAA81917',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Anyone But You (2023) [720p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/anyone-but-you-2023"><img src="https://img.yts.mx/assets/images/movies/anyone_but_you_2023/medium-cover.jpg" alt="Anyone But You (2023)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Comedy / Romance<br />Size: 951.22 MB<br />Runtime: 1hr 43 min<br /><br />In the aftermath of a captivating first encounter, Bea and Ben discover that their initial fiery passion has inexplicably dwindled to frosty misunderstanding. However, fate intervenes: Thrown together again at a dreamy Australian wedding, they opt for a charade of coupledom. But sparks reignite amid the sun-kissed scenery, forcing them to confront their true feelings and embrace a second chance at love.—Raihan-Khondker',
+          link: 'https://yts.mx/movies/anyone-but-you-2023',
+          guid: 'https://yts.mx/movies/anyone-but-you-2023#720p',
+          pubDate: 'Sun, 18 Feb 2024 09:35:54 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/478AA2D5DC68C6AE4B276470C30CE6F328DAB53E',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Anyone But You (2023) [1080p] [WEBRip] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/anyone-but-you-2023"><img src="https://img.yts.mx/assets/images/movies/anyone_but_you_2023/medium-cover.jpg" alt="Anyone But You (2023)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Comedy / Romance<br />Size: 1.91 GB<br />Runtime: 1hr 43 min<br /><br />In the aftermath of a captivating first encounter, Bea and Ben discover that their initial fiery passion has inexplicably dwindled to frosty misunderstanding. However, fate intervenes: Thrown together again at a dreamy Australian wedding, they opt for a charade of coupledom. But sparks reignite amid the sun-kissed scenery, forcing them to confront their true feelings and embrace a second chance at love.—Raihan-Khondker',
+          link: 'https://yts.mx/movies/anyone-but-you-2023',
+          guid: 'https://yts.mx/movies/anyone-but-you-2023#1080p',
+          pubDate: 'Sun, 18 Feb 2024 10:36:17 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/DF70A15A89CEDE0957A0C4DF685DF08A4F9D680D',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title:
+            'Anyone But You (2023) [1080p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/anyone-but-you-2023"><img src="https://img.yts.mx/assets/images/movies/anyone_but_you_2023/medium-cover.jpg" alt="Anyone But You (2023)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Comedy / Romance<br />Size: 1.73 GB<br />Runtime: 1hr 43 min<br /><br />In the aftermath of a captivating first encounter, Bea and Ben discover that their initial fiery passion has inexplicably dwindled to frosty misunderstanding. However, fate intervenes: Thrown together again at a dreamy Australian wedding, they opt for a charade of coupledom. But sparks reignite amid the sun-kissed scenery, forcing them to confront their true feelings and embrace a second chance at love.—Raihan-Khondker',
+          link: 'https://yts.mx/movies/anyone-but-you-2023',
+          guid: 'https://yts.mx/movies/anyone-but-you-2023#1080p',
+          pubDate: 'Sun, 18 Feb 2024 14:40:08 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/CC089FF26DE6AC5E99ECE4F716F58892BA8EBF89',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Cause for Alarm! (1951) [720p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/cause-for-alarm-1951"><img src="https://img.yts.mx/assets/images/movies/cause_for_alarm_1951/medium-cover.jpg" alt="Cause for Alarm! (1951)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Crime / Drama / Film-Noir / Thriller<br />Size: 686.23 MB<br />Runtime: 1hr 14 min<br /><br />Invalid George Jones (Barry Sullivan) is both physically and mentally ill. He mistakenly believes his wife Ellen (Loretta Young) and his doctor (Bruce Cowling) are having an affair and also planning to kill him. He writes a letter to the district attorney detailing their alleged murder plot. After he has Ellen give the letter to their postman, he reveals its contents to her and then threatens her with a gun. The excitement proves to much and George suffers a fatal collapse. Now Ellen must find a way to retrieve the incriminating letter.—Daniel Bubbeo',
+          link: 'https://yts.mx/movies/cause-for-alarm-1951',
+          guid: 'https://yts.mx/movies/cause-for-alarm-1951#720p',
+          pubDate: 'Sun, 18 Feb 2024 13:48:53 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/CB7E1F4733948A50D974A890950EACF03395ED35',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Cause for Alarm! (1951) [1080p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/cause-for-alarm-1951"><img src="https://img.yts.mx/assets/images/movies/cause_for_alarm_1951/medium-cover.jpg" alt="Cause for Alarm! (1951)" /></a><br />IMDB Rating: 6.4/10<br />Genre: Crime / Drama / Film-Noir / Thriller<br />Size: 1.24 GB<br />Runtime: 1hr 14 min<br /><br />Invalid George Jones (Barry Sullivan) is both physically and mentally ill. He mistakenly believes his wife Ellen (Loretta Young) and his doctor (Bruce Cowling) are having an affair and also planning to kill him. He writes a letter to the district attorney detailing their alleged murder plot. After he has Ellen give the letter to their postman, he reveals its contents to her and then threatens her with a gun. The excitement proves to much and George suffers a fatal collapse. Now Ellen must find a way to retrieve the incriminating letter.—Daniel Bubbeo',
+          link: 'https://yts.mx/movies/cause-for-alarm-1951',
+          guid: 'https://yts.mx/movies/cause-for-alarm-1951#1080p',
+          pubDate: 'Sun, 18 Feb 2024 14:36:31 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/4163A72C080E7C00B8AC175FBA795F684EEA261E',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Tremors (1990) [720p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/tremors-1990"><img src="https://img.yts.mx/assets/images/movies/Tremors_1990/medium-cover.jpg" alt="Tremors (1990)" /></a><br />IMDB Rating: 7.1/10<br />Genre: Action / Comedy / Horror / Sci-Fi<br />Size: 699.56 MB<br />Runtime: 1hr 36 min<br /><br />Trying to escape their dead-end life in the desert town of Perfection, Nevada, Valentine "Val" McKee and Earl Bassett find themselves sidetracked when corpses mysteriously start piling up around them. When Val and Earl have a run-in with some creepy tentacled creatures that have made lunch out of a road crew, they retreat back to the town to spread the news. Although these monsters are not exactly intellectuals, their mental powers are still mind-boggling. Detecting a human hiding in a car, they dig under the vehicle, causing it to sink into the earth. Val and Earl get help from several townspeople. With the help of Rhonda, Val and Earl kill one of the monsters. But there are three of them left, and they are each about 30 feet long. Walter starts calling the monsters graboids. When the survivors think they have outwitted the graboids by taking refuge atop the roofs of buildings, the graboids merely destroy the buildings\' foundations, killing two people. The graboids are quick learners when it comes to stalking their prey, and the humans must continually be on their toes if they are to stay alive while they try to find a way to defeat the graboids.—Todd Baldridge',
+          link: 'https://yts.mx/movies/tremors-1990',
+          guid: 'https://yts.mx/movies/tremors-1990#720p',
+          pubDate: 'Sun, 01 Nov 2015 03:47:07 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/5372C63146C6BDD8950E4D47FAC6703C10714939',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Tremors (1990) [1080p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/tremors-1990"><img src="https://img.yts.mx/assets/images/movies/Tremors_1990/medium-cover.jpg" alt="Tremors (1990)" /></a><br />IMDB Rating: 7.1/10<br />Genre: Action / Comedy / Horror / Sci-Fi<br />Size: 1.30 GB<br />Runtime: 1hr 36 min<br /><br />Trying to escape their dead-end life in the desert town of Perfection, Nevada, Valentine "Val" McKee and Earl Bassett find themselves sidetracked when corpses mysteriously start piling up around them. When Val and Earl have a run-in with some creepy tentacled creatures that have made lunch out of a road crew, they retreat back to the town to spread the news. Although these monsters are not exactly intellectuals, their mental powers are still mind-boggling. Detecting a human hiding in a car, they dig under the vehicle, causing it to sink into the earth. Val and Earl get help from several townspeople. With the help of Rhonda, Val and Earl kill one of the monsters. But there are three of them left, and they are each about 30 feet long. Walter starts calling the monsters graboids. When the survivors think they have outwitted the graboids by taking refuge atop the roofs of buildings, the graboids merely destroy the buildings\' foundations, killing two people. The graboids are quick learners when it comes to stalking their prey, and the humans must continually be on their toes if they are to stay alive while they try to find a way to defeat the graboids.—Todd Baldridge',
+          link: 'https://yts.mx/movies/tremors-1990',
+          guid: 'https://yts.mx/movies/tremors-1990#1080p',
+          pubDate: 'Sun, 01 Nov 2015 03:47:09 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/4FC917FFD092801AFF50FE1E568ED7D0504E336C',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title:
+            'Tremors (1990) [2160p] [BluRay] [x265] [10bit] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/tremors-1990"><img src="https://img.yts.mx/assets/images/movies/Tremors_1990/medium-cover.jpg" alt="Tremors (1990)" /></a><br />IMDB Rating: 7.1/10<br />Genre: Action / Comedy / Horror / Sci-Fi<br />Size: 4.28 GB<br />Runtime: 1hr 36 min<br /><br />Trying to escape their dead-end life in the desert town of Perfection, Nevada, Valentine "Val" McKee and Earl Bassett find themselves sidetracked when corpses mysteriously start piling up around them. When Val and Earl have a run-in with some creepy tentacled creatures that have made lunch out of a road crew, they retreat back to the town to spread the news. Although these monsters are not exactly intellectuals, their mental powers are still mind-boggling. Detecting a human hiding in a car, they dig under the vehicle, causing it to sink into the earth. Val and Earl get help from several townspeople. With the help of Rhonda, Val and Earl kill one of the monsters. But there are three of them left, and they are each about 30 feet long. Walter starts calling the monsters graboids. When the survivors think they have outwitted the graboids by taking refuge atop the roofs of buildings, the graboids merely destroy the buildings\' foundations, killing two people. The graboids are quick learners when it comes to stalking their prey, and the humans must continually be on their toes if they are to stay alive while they try to find a way to defeat the graboids.—Todd Baldridge',
+          link: 'https://yts.mx/movies/tremors-1990',
+          guid: 'https://yts.mx/movies/tremors-1990#2160p',
+          pubDate: 'Sun, 18 Feb 2024 14:00:27 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/7E05139707F19A0222B5C5EB0F89BD3B18FACFBB',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'The Sidemen Story (2024) [720p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/the-sidemen-story-2024"><img src="https://img.yts.mx/assets/images/movies/the_sidemen_story_2024/medium-cover.jpg" alt="The Sidemen Story (2024)" /></a><br />IMDB Rating: 6.9/10<br />Genre: Biography / Documentary<br />Size: 917.58 MB<br />Runtime: 1hr 39 min<br /><br />',
+          link: 'https://yts.mx/movies/the-sidemen-story-2024',
+          guid: 'https://yts.mx/movies/the-sidemen-story-2024#720p',
+          pubDate: 'Sun, 18 Feb 2024 09:36:57 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/7FC82D356B278C41D345D1C77B28061FE7256F69',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'The Sidemen Story (2024) [1080p] [WEBRip] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/the-sidemen-story-2024"><img src="https://img.yts.mx/assets/images/movies/the_sidemen_story_2024/medium-cover.jpg" alt="The Sidemen Story (2024)" /></a><br />IMDB Rating: 6.9/10<br />Genre: Biography / Documentary<br />Size: 1.84 GB<br />Runtime: 1hr 39 min<br /><br />',
+          link: 'https://yts.mx/movies/the-sidemen-story-2024',
+          guid: 'https://yts.mx/movies/the-sidemen-story-2024#1080p',
+          pubDate: 'Sun, 18 Feb 2024 10:58:40 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/37BC252B31034ECCCC8B2F5AB1454CF54FEB33B6',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Nothing But a Man (1964) [720p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/nothing-but-a-man-1964"><img src="https://img.yts.mx/assets/images/movies/nothing_but_a_man_1964/medium-cover.jpg" alt="Nothing But a Man (1964)" /></a><br />IMDB Rating: 7.9/10<br />Genre: Action / Drama / Romance<br />Size: 841.39 MB<br />Runtime: 1hr 35 min<br /><br />Born in Birmingham, Duff Anderson, the father of a male toddler, who lives with a nanny, re-locates to a small town to work on the railroad. He meets with and is attracted to Josie much to the chagrin of her preacher father. The marriage does take place nevertheless, both re-locate to live in their own house and he gets a job in a mill. He decides not to bring his son to live with them. Challenges arise when the Mill Foreman finds out that Duff is attempting to unionize the workers, forcing Duff to quit, and look for work elsewhere. Unable to reconcile himself to working on a daily wage of $2.50 picking cotton nor even as a waiter, he gets a job at a garage. He is enraged at a customer for belittling him and Josie, and is let go. Unemployed, unable to support his wife and son, he gets abusive and leaves - perhaps never to return.—rAjOo (gunwanti@hotmail.com)',
+          link: 'https://yts.mx/movies/nothing-but-a-man-1964',
+          guid: 'https://yts.mx/movies/nothing-but-a-man-1964#720p',
+          pubDate: 'Sun, 18 Feb 2024 09:47:56 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/4DBA9B04093E08A319BC368E2270489560EFEBA0',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Nothing But a Man (1964) [1080p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/nothing-but-a-man-1964"><img src="https://img.yts.mx/assets/images/movies/nothing_but_a_man_1964/medium-cover.jpg" alt="Nothing But a Man (1964)" /></a><br />IMDB Rating: 7.9/10<br />Genre: Action / Drama / Romance<br />Size: 1.53 GB<br />Runtime: 1hr 35 min<br /><br />Born in Birmingham, Duff Anderson, the father of a male toddler, who lives with a nanny, re-locates to a small town to work on the railroad. He meets with and is attracted to Josie much to the chagrin of her preacher father. The marriage does take place nevertheless, both re-locate to live in their own house and he gets a job in a mill. He decides not to bring his son to live with them. Challenges arise when the Mill Foreman finds out that Duff is attempting to unionize the workers, forcing Duff to quit, and look for work elsewhere. Unable to reconcile himself to working on a daily wage of $2.50 picking cotton nor even as a waiter, he gets a job at a garage. He is enraged at a customer for belittling him and Josie, and is let go. Unemployed, unable to support his wife and son, he gets abusive and leaves - perhaps never to return.—rAjOo (gunwanti@hotmail.com)',
+          link: 'https://yts.mx/movies/nothing-but-a-man-1964',
+          guid: 'https://yts.mx/movies/nothing-but-a-man-1964#1080p',
+          pubDate: 'Sun, 18 Feb 2024 10:50:53 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/B49A191BF5E983B49BD7CFF6C73514F3964A5033',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Nothing But a Man (1964) [720p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/nothing-but-a-man-1964"><img src="https://img.yts.mx/assets/images/movies/nothing_but_a_man_1964/medium-cover.jpg" alt="Nothing But a Man (1964)" /></a><br />IMDB Rating: 7.9/10<br />Genre: Action / Drama / Romance<br />Size: 845.53 MB<br />Runtime: 1hr 35 min<br /><br />Born in Birmingham, Duff Anderson, the father of a male toddler, who lives with a nanny, re-locates to a small town to work on the railroad. He meets with and is attracted to Josie much to the chagrin of her preacher father. The marriage does take place nevertheless, both re-locate to live in their own house and he gets a job in a mill. He decides not to bring his son to live with them. Challenges arise when the Mill Foreman finds out that Duff is attempting to unionize the workers, forcing Duff to quit, and look for work elsewhere. Unable to reconcile himself to working on a daily wage of $2.50 picking cotton nor even as a waiter, he gets a job at a garage. He is enraged at a customer for belittling him and Josie, and is let go. Unemployed, unable to support his wife and son, he gets abusive and leaves - perhaps never to return.—rAjOo (gunwanti@hotmail.com)',
+          link: 'https://yts.mx/movies/nothing-but-a-man-1964',
+          guid: 'https://yts.mx/movies/nothing-but-a-man-1964#720p',
+          pubDate: 'Fri, 27 May 2022 21:10:12 +0200',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/DA9CCA19B44249B542BB6F6C973FD14EB82E7133',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Nothing But a Man (1964) [1080p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/nothing-but-a-man-1964"><img src="https://img.yts.mx/assets/images/movies/nothing_but_a_man_1964/medium-cover.jpg" alt="Nothing But a Man (1964)" /></a><br />IMDB Rating: 7.9/10<br />Genre: Action / Drama / Romance<br />Size: 1.53 GB<br />Runtime: 1hr 35 min<br /><br />Born in Birmingham, Duff Anderson, the father of a male toddler, who lives with a nanny, re-locates to a small town to work on the railroad. He meets with and is attracted to Josie much to the chagrin of her preacher father. The marriage does take place nevertheless, both re-locate to live in their own house and he gets a job in a mill. He decides not to bring his son to live with them. Challenges arise when the Mill Foreman finds out that Duff is attempting to unionize the workers, forcing Duff to quit, and look for work elsewhere. Unable to reconcile himself to working on a daily wage of $2.50 picking cotton nor even as a waiter, he gets a job at a garage. He is enraged at a customer for belittling him and Josie, and is let go. Unemployed, unable to support his wife and son, he gets abusive and leaves - perhaps never to return.—rAjOo (gunwanti@hotmail.com)',
+          link: 'https://yts.mx/movies/nothing-but-a-man-1964',
+          guid: 'https://yts.mx/movies/nothing-but-a-man-1964#1080p',
+          pubDate: 'Fri, 27 May 2022 22:44:36 +0200',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/BA6BB2AD15868C080C98064FBA745FC69ABDD7CC',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Story Ave (2023) [720p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/story-ave-2023"><img src="https://img.yts.mx/assets/images/movies/story_ave_2023/medium-cover.jpg" alt="Story Ave (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama<br />Size: 979.08 MB<br />Runtime: 1hr 46 min<br /><br />',
+          link: 'https://yts.mx/movies/story-ave-2023',
+          guid: 'https://yts.mx/movies/story-ave-2023#720p',
+          pubDate: 'Sun, 18 Feb 2024 07:48:59 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/FEBB30242157C8EA08B09B98FE3EFF98486ACFEF',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Story Ave (2023) [1080p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/story-ave-2023"><img src="https://img.yts.mx/assets/images/movies/story_ave_2023/medium-cover.jpg" alt="Story Ave (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama<br />Size: 1.77 GB<br />Runtime: 1hr 46 min<br /><br />',
+          link: 'https://yts.mx/movies/story-ave-2023',
+          guid: 'https://yts.mx/movies/story-ave-2023#1080p',
+          pubDate: 'Sun, 18 Feb 2024 09:02:35 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/BCC54D6FE8AE4DB22F87B3743E9D12DC44086F4F',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Story Ave (2023) [720p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/story-ave-2023"><img src="https://img.yts.mx/assets/images/movies/story_ave_2023/medium-cover.jpg" alt="Story Ave (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama<br />Size: 979.48 MB<br />Runtime: 1hr 46 min<br /><br />',
+          link: 'https://yts.mx/movies/story-ave-2023',
+          guid: 'https://yts.mx/movies/story-ave-2023#720p',
+          pubDate: 'Tue, 06 Feb 2024 08:55:20 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/D46BCC209A3617639AB631DC5EA92E8A12FB368E',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Story Ave (2023) [1080p] [WEBRip] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/story-ave-2023"><img src="https://img.yts.mx/assets/images/movies/story_ave_2023/medium-cover.jpg" alt="Story Ave (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama<br />Size: 1.97 GB<br />Runtime: 1hr 46 min<br /><br />',
+          link: 'https://yts.mx/movies/story-ave-2023',
+          guid: 'https://yts.mx/movies/story-ave-2023#1080p',
+          pubDate: 'Tue, 06 Feb 2024 10:00:46 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/618E90EE3C619B74053007A47A6690CAF0A7654E',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Eileen (2023) [720p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 895.96 MB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
+          link: 'https://yts.mx/movies/eileen-2023',
+          guid: 'https://yts.mx/movies/eileen-2023#720p',
+          pubDate: 'Sun, 18 Feb 2024 07:45:09 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/172D82D76C74087363E5D164D921E2ACCE0C979A',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Eileen (2023) [1080p] [BluRay] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 1.8 GB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
+          link: 'https://yts.mx/movies/eileen-2023',
+          guid: 'https://yts.mx/movies/eileen-2023#1080p',
+          pubDate: 'Sun, 18 Feb 2024 09:01:01 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/49C22DB047A202F87C8AF313452F9A12466F79B4',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Eileen (2023) [720p] [WEBRip] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 896.54 MB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
+          link: 'https://yts.mx/movies/eileen-2023',
+          guid: 'https://yts.mx/movies/eileen-2023#720p',
+          pubDate: 'Tue, 02 Jan 2024 07:04:08 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/0F91C0492E41E08CBB76810D484FB105DBF91D8A',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Eileen (2023) [1080p] [WEBRip] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 1.8 GB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
+          link: 'https://yts.mx/movies/eileen-2023',
+          guid: 'https://yts.mx/movies/eileen-2023#1080p',
+          pubDate: 'Tue, 02 Jan 2024 08:15:25 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/D7AE2DC467B381F5C8A86FF91F1846B722C8E3E0',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Eileen (2023) [1080p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 1.63 GB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
+          link: 'https://yts.mx/movies/eileen-2023',
+          guid: 'https://yts.mx/movies/eileen-2023#1080p',
+          pubDate: 'Tue, 02 Jan 2024 13:47:24 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/0E7F3409C048AFA061915905556B1E6F2F85ACE8',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Eileen (2023) [2160p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/eileen-2023"><img src="https://img.yts.mx/assets/images/movies/eileen_2023/medium-cover.jpg" alt="Eileen (2023)" /></a><br />IMDB Rating: 6.1/10<br />Genre: Action / Drama / Mystery / Thriller<br />Size: 4.34 GB<br />Runtime: 1hr 37 min<br /><br />The stagnant waters of Eileen\'s dull, stifled life as a solitary worker at a juvenile detention center in 1960s Boston, are unexpectedly disrupted when the institution brings in a new psychologist, the vibrant Rebecca. The fervent enthusiasm that blossoms between the two women almost immediately gives way to a closer relationship, until their fragile connection takes a dramatic turn.—Thessaloniki International Film Festival',
+          link: 'https://yts.mx/movies/eileen-2023',
+          guid: 'https://yts.mx/movies/eileen-2023#2160p',
+          pubDate: 'Thu, 04 Jan 2024 00:28:04 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/CA6121F4BF489BC13E57152EE1A3187360D2E57C',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'The Selling (2011) [720p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/the-selling-2011"><img src="https://img.yts.mx/assets/images/movies/the_selling_2011/medium-cover.jpg" alt="The Selling (2011)" /></a><br />IMDB Rating: 6.0/10<br />Genre: Comedy<br />Size: 826.11 MB<br />Runtime: 1hr 32 min<br /><br />',
+          link: 'https://yts.mx/movies/the-selling-2011',
+          guid: 'https://yts.mx/movies/the-selling-2011#720p',
+          pubDate: 'Sun, 18 Feb 2024 05:14:06 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/725A1D7B2BCA4FDD7485021696A672B8ACFE5AD2',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'The Selling (2011) [1080p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/the-selling-2011"><img src="https://img.yts.mx/assets/images/movies/the_selling_2011/medium-cover.jpg" alt="The Selling (2011)" /></a><br />IMDB Rating: 6.0/10<br />Genre: Comedy<br />Size: 1.5 GB<br />Runtime: 1hr 32 min<br /><br />',
+          link: 'https://yts.mx/movies/the-selling-2011',
+          guid: 'https://yts.mx/movies/the-selling-2011#1080p',
+          pubDate: 'Sun, 18 Feb 2024 06:14:27 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/228A9F41684B55E34CECEE2A8151374A35B5FD66',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Repligator (1998) [720p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/repligator-1998"><img src="https://img.yts.mx/assets/images/movies/repligator_1998/medium-cover.jpg" alt="Repligator (1998)" /></a><br />IMDB Rating: 2.6/10<br />Genre: Action / Comedy / Sci-Fi<br />Size: 786.83 MB<br />Runtime: 1hr 26 min<br /><br />',
+          link: 'https://yts.mx/movies/repligator-1998',
+          guid: 'https://yts.mx/movies/repligator-1998#720p',
+          pubDate: 'Sun, 18 Feb 2024 03:39:54 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/C101922D0BD1A7E1CEB544142E780969CFE2687E',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Repligator (1998) [1080p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/repligator-1998"><img src="https://img.yts.mx/assets/images/movies/repligator_1998/medium-cover.jpg" alt="Repligator (1998)" /></a><br />IMDB Rating: 2.6/10<br />Genre: Action / Comedy / Sci-Fi<br />Size: 1.42 GB<br />Runtime: 1hr 26 min<br /><br />',
+          link: 'https://yts.mx/movies/repligator-1998',
+          guid: 'https://yts.mx/movies/repligator-1998#1080p',
+          pubDate: 'Sun, 18 Feb 2024 04:39:20 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/AB7D6B62CC97CA52F53B912FEF25823F43BB53B7',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'This Is Me... Now (2024) [720p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/this-is-me-now-2024"><img src="https://img.yts.mx/assets/images/movies/this_is_me_now_2024/medium-cover.jpg" alt="This Is Me... Now (2024)" /></a><br />IMDB Rating: 5.3/10<br />Genre: Music / Musical<br />Size: 603 MB<br />Runtime: 1hr 5 min<br /><br />',
+          link: 'https://yts.mx/movies/this-is-me-now-2024',
+          guid: 'https://yts.mx/movies/this-is-me-now-2024#720p',
+          pubDate: 'Fri, 16 Feb 2024 01:31:38 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/1026B9FA4E75F11569917E7BF56D7005DAE68CA6',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'This Is Me... Now (2024) [1080p] [WEBRip] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/this-is-me-now-2024"><img src="https://img.yts.mx/assets/images/movies/this_is_me_now_2024/medium-cover.jpg" alt="This Is Me... Now (2024)" /></a><br />IMDB Rating: 5.3/10<br />Genre: Music / Musical<br />Size: 1.21 GB<br />Runtime: 1hr 5 min<br /><br />',
+          link: 'https://yts.mx/movies/this-is-me-now-2024',
+          guid: 'https://yts.mx/movies/this-is-me-now-2024#1080p',
+          pubDate: 'Fri, 16 Feb 2024 02:08:39 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/29452026DB59FB985AC42C9219D066394E1EC04A',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title:
+            'This Is Me... Now (2024) [2160p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/this-is-me-now-2024"><img src="https://img.yts.mx/assets/images/movies/this_is_me_now_2024/medium-cover.jpg" alt="This Is Me... Now (2024)" /></a><br />IMDB Rating: 5.3/10<br />Genre: Music / Musical<br />Size: 2.92 GB<br />Runtime: 1hr 5 min<br /><br />',
+          link: 'https://yts.mx/movies/this-is-me-now-2024',
+          guid: 'https://yts.mx/movies/this-is-me-now-2024#2160p',
+          pubDate: 'Sun, 18 Feb 2024 04:16:08 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/7FE6D31CC25E999D6A09305786D43BA6B7C5F57E',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Has Anybody Seen My Gal (1952) [720p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/has-anybody-seen-my-gal-1952"><img src="https://img.yts.mx/assets/images/movies/has_anybody_seen_my_gal_1952/medium-cover.jpg" alt="Has Anybody Seen My Gal (1952)" /></a><br />IMDB Rating: 7.1/10<br />Genre: Comedy<br />Size: 816.11 MB<br />Runtime: 1hr 28 min<br /><br />Wealthy Samuel Fulton is getting older and has no family of his own. He decides to leave his estate to the family of his first love, who turned down his marriage proposal years ago because he was poor. But he wants to test the family before leaving his money to them. He takes a room in their home and a job in the father\'s shop. He anonymously grants them $100,000. Harriet Blaisdell moves the family into a mansion and makes plans to marry her daughter Millicent off to a socialite rather than her soda jerk boyfriend Dan. The money goes to their heads, and they soon find themselves broke, back in their old house, and back to their old lives. Father back in his shop, Millicent engaged to Dan, and everyone seemingly much happier. Hoping they learned their lesson, Fulton takes his leave of the family.—Ed Stephan',
+          link: 'https://yts.mx/movies/has-anybody-seen-my-gal-1952',
+          guid: 'https://yts.mx/movies/has-anybody-seen-my-gal-1952#720p',
+          pubDate: 'Sat, 17 Feb 2024 22:58:56 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/74FF3CE61FFE874AA862941229BD6DE708B22940',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: 'Has Anybody Seen My Gal (1952) [1080p] [BluRay] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/has-anybody-seen-my-gal-1952"><img src="https://img.yts.mx/assets/images/movies/has_anybody_seen_my_gal_1952/medium-cover.jpg" alt="Has Anybody Seen My Gal (1952)" /></a><br />IMDB Rating: 7.1/10<br />Genre: Comedy<br />Size: 1.48 GB<br />Runtime: 1hr 28 min<br /><br />Wealthy Samuel Fulton is getting older and has no family of his own. He decides to leave his estate to the family of his first love, who turned down his marriage proposal years ago because he was poor. But he wants to test the family before leaving his money to them. He takes a room in their home and a job in the father\'s shop. He anonymously grants them $100,000. Harriet Blaisdell moves the family into a mansion and makes plans to marry her daughter Millicent off to a socialite rather than her soda jerk boyfriend Dan. The money goes to their heads, and they soon find themselves broke, back in their old house, and back to their old lives. Father back in his shop, Millicent engaged to Dan, and everyone seemingly much happier. Hoping they learned their lesson, Fulton takes his leave of the family.—Ed Stephan',
+          link: 'https://yts.mx/movies/has-anybody-seen-my-gal-1952',
+          guid: 'https://yts.mx/movies/has-anybody-seen-my-gal-1952#1080p',
+          pubDate: 'Sun, 18 Feb 2024 00:00:35 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/854ECFF72F7434E7F165CE6B43BD6B25EB550A5E',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title:
+            'Snoopy Presents: Welcome Home, Franklin (2024) [720p] [WEBRip] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024"><img src="https://img.yts.mx/assets/images/movies/snoopy_presents_welcome_home_franklin_2024/medium-cover.jpg" alt="Snoopy Presents: Welcome Home, Franklin (2024)" /></a><br />IMDB Rating: 7.2/10<br />Genre: Animation<br />Size: 360.67 MB<br />Runtime: 12hr 39 min<br /><br />The origin story for one of Peanuts\' most beloved characters, Franklin, follows how he approaches making new friends. Franklin\'s family is always on the move with his dad\'s military job, and everywhere he goes Franklin finds support in a notebook filled with his grandfather\'s advice on friendship. But when Franklin tries his usual strategies with the Peanuts gang, he has trouble fitting in. That\'s until he learns about the neighborhood Soap Box Derby race - according to his grandfather, everyone loves a winner. He\'s sure that winning the race will also mean winning over some new friends. All he needs is a partner, which he finds in Charlie Brown. Franklin and Charlie Brown work together to build a car and in the process become good buddies. But as the race nears, the pressure mounts - can their car and their newfound friendship make it to the finish line?—Apple TV+',
+          link: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024',
+          guid: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024#720p',
+          pubDate: 'Fri, 16 Feb 2024 03:35:46 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/6857C0A27586A351E9B404756CB77A39B541E51F',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title:
+            'Snoopy Presents: Welcome Home, Franklin (2024) [1080p] [WEBRip] [5.1] [YTS.MX]',
+          description:
+            '<a href="https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024"><img src="https://img.yts.mx/assets/images/movies/snoopy_presents_welcome_home_franklin_2024/medium-cover.jpg" alt="Snoopy Presents: Welcome Home, Franklin (2024)" /></a><br />IMDB Rating: 7.2/10<br />Genre: Animation<br />Size: 738.44 MB<br />Runtime: 12hr 39 min<br /><br />The origin story for one of Peanuts\' most beloved characters, Franklin, follows how he approaches making new friends. Franklin\'s family is always on the move with his dad\'s military job, and everywhere he goes Franklin finds support in a notebook filled with his grandfather\'s advice on friendship. But when Franklin tries his usual strategies with the Peanuts gang, he has trouble fitting in. That\'s until he learns about the neighborhood Soap Box Derby race - according to his grandfather, everyone loves a winner. He\'s sure that winning the race will also mean winning over some new friends. All he needs is a partner, which he finds in Charlie Brown. Franklin and Charlie Brown work together to build a car and in the process become good buddies. But as the race nears, the pressure mounts - can their car and their newfound friendship make it to the finish line?—Apple TV+',
+          link: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024',
+          guid: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024#1080p',
+          pubDate: 'Fri, 16 Feb 2024 03:56:42 +0100',
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/3F5DDC35EF8B304C7AFEC7EF8F8F7C44FAD68EAA',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+        {
+          title: `Snoopy Presents: Welcome Home, Franklin (${new Date().getFullYear()}) [2160p] [WEBRip] [x265] [10bit] [5.1] [YTS.MX]`,
+          description:
+            '<a href="https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024"><img src="https://img.yts.mx/assets/images/movies/snoopy_presents_welcome_home_franklin_2024/medium-cover.jpg" alt="Snoopy Presents: Welcome Home, Franklin (2024)" /></a><br />IMDB Rating: 7.2/10<br />Genre: Animation<br />Size: 1.74 GB<br />Runtime: 12hr 39 min<br /><br />The origin story for one of Peanuts\' most beloved characters, Franklin, follows how he approaches making new friends. Franklin\'s family is always on the move with his dad\'s military job, and everywhere he goes Franklin finds support in a notebook filled with his grandfather\'s advice on friendship. But when Franklin tries his usual strategies with the Peanuts gang, he has trouble fitting in. That\'s until he learns about the neighborhood Soap Box Derby race - according to his grandfather, everyone loves a winner. He\'s sure that winning the race will also mean winning over some new friends. All he needs is a partner, which he finds in Charlie Brown. Franklin and Charlie Brown work together to build a car and in the process become good buddies. But as the race nears, the pressure mounts - can their car and their newfound friendship make it to the finish line?—Apple TV+',
+          link: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024',
+          guid: 'https://yts.mx/movies/snoopy-presents-welcome-home-franklin-2024#2160p',
+          pubDate: String(new Date()),
+          enclosure: {
+            url: 'https://yts.mx/torrent/download/22987BCBE6A58AC7AE393B5326604D047F096273',
+            type: 'application/x-bittorrent',
+            length: '10000',
+          },
+        },
+      ],
+    },
+    version: '2.0',
+    'xmlns:dc': 'http://purl.org/dc/elements/1.1/',
+    'xmlns:content': 'http://purl.org/rss/1.0/modules/content/',
+    'xmlns:atom': 'http://www.w3.org/2005/Atom',
+  },
+};
